Divergence Index

Elizabeth Roberto's Divergence index for calculating and decomposing segregation.

divergence(
  ...,
  population = NA,
  na.rm = TRUE,
  summed = FALSE,
  logBase = exp(1),
  comparison = NULL,
  sumPercent = NA,
  weights = NA,
  rowTotals = NA
)

Arguments

...
Vectors containing proportions of each group, or a dataframe containing only those vectors.

population
A vector of population totals for each row, or weights summing to a total of one. If NA, will assume all populations/weights are equal. Set to 1 to silence warning. If a string, will use the string as the named column of the dataframe provided in ...

na.rm
logical. Should missing values (including NaN) be removed? Used only if summed is set to TRUE.

summed
If TRUE, will return a single summary statistic. (Or one value per group if specifying dplyr::group_by.) If FALSE (default), will return a vector equaling the length of the input vectors. If 'weighted' (only for divergence and information theory), returns a vector as in FALSE, but with pre-weighted values, such that sum(divergence(..., summed = 'weighted)) is equivalent to divergence(..., summed = T).

logBase
Specify the base for the logathirm used in the equation. Natural logarithm by default.

comparison
A vector of percentages that must equal the length of the number of vectors or columns in ..., representing the percentages of each group in the larger (comparison) geography.

weights
deprecated, use population.

rowTotals, sumPercent
deprecated, will throw error.

Source

Created by Elizabeth Roberto: https://arxiv.org/abs/1508.01167

Value

A single value if summed==TRUE, or a vector equaling the length of the inputs.

Details

The demographics of each observation are compared to that of their larger geography, which is inferred from the combination of the percentages in ... and the population totals in population, or can be directly provided with comparison.

Examples

library(rsegregation)
data("bay_race")
#return by-observation scores
divergence(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, bay_race$all_other)

#> Warning: Population parameter not set; assuming equal populations.

#>    [1] 0.206899484 0.278874430 0.182497789 0.191584401 0.310336077 0.522469045
#>    [7] 0.701664418 0.481644000 0.580122593 0.719875959 0.197358065 0.182535530
#>   [13] 0.365182118 0.704893826 0.790448748 0.641153838 0.351999538 0.979223477
#>   [19] 0.475718280 0.950025068 1.259108921 0.604057877 0.898702071 0.516255716
#>   [25] 0.269706079 1.050711188 0.360805121 0.310377577 0.171313004 0.254979490
#>   [31] 0.235728300 0.427752109 0.175822306 0.189313703 0.126982166 0.138399554
#>   [37] 0.126194624 0.142816556 0.162316085 0.214988468 0.222143025 0.154401886
#>   [43] 0.209676454 0.217446818 0.221382017 0.213057459 0.128782060 0.076311677
#>   [49] 0.156525564 0.222527898 0.113612697 0.161306964 0.197944586 0.330698839
#>   [55] 0.432116159 0.340226114 0.263726785 0.471283910 0.493883426 0.468663608
#>   [61] 0.457391162 0.237312238 0.346870697 0.508270457 0.577129818 0.397580247
#>   [67] 0.239664753 0.378010619 0.353232476 0.352569831 0.091803604 0.089158580
#>   [73] 0.145684943 0.334045051 0.468487327 0.425171220 0.680904224 0.476968349
#>   [79] 0.719662744 0.851754833 0.535226971 0.735350035 0.325122226 0.155547632
#>   [85] 0.085364015 0.297666593 0.839563653 0.595863334 1.094420848 1.077119427
#>   [91] 1.016660436 0.941177808 0.919338534 0.949734528 1.045219482 1.052147561
#>   [97] 0.970718014 0.924293212 0.866610436 0.941604292 0.967447888 1.045513813
#>  [103] 1.169044309 0.795615187 0.692329215 1.095770968 1.058664132 0.903396461
#>  [109] 0.815002957 1.113775634 0.115837374 0.078643130 0.106971334 0.185250534
#>  [115] 0.059346147 0.167726523 0.261606672 0.299937640 0.285481524 0.346921475
#>  [121] 0.305344192 0.275183558 0.158688727 0.256563846 0.115873436 0.260433351
#>  [127] 0.089345939 0.084619687 0.091399807 0.095805434 0.128960057 0.125161333
#>  [133] 0.086985245 0.143464723 0.089975184 0.087746385 0.097523317 0.201083069
#>  [139] 0.354094079 0.204602655 0.156944205 0.163197499 0.138565169 0.145949498
#>  [145] 0.371669714 0.214590128 0.199686174 0.307836118 0.520089220 0.159329928
#>  [151] 0.162057715 0.112440774 0.184354711 0.254838142 0.213537472 0.153384145
#>  [157] 0.057365015 0.071225049 0.190664273 0.049488194 0.055293384 0.055160244
#>  [163] 0.045668293 0.066163086 0.077256796 0.158856968 0.195199611 0.033515733
#>  [169] 0.078023081 0.111702315 0.101156722 0.180472862 0.132362503 0.098708795
#>  [175] 0.108017081 0.115987632 0.124072737 0.039269654 0.056252869 0.041475629
#>  [181] 0.030857025 0.009380802 0.075208343 0.052114813 0.061354502 0.135099827
#>  [187] 0.121650820 0.102874708 0.151274025 0.220501879 0.137487930 0.021762361
#>  [193] 0.020075368 0.021340131 0.123755185 0.332846593 0.207072910 0.121188834
#>  [199] 0.104913017 0.252114828 0.066267347 0.100690685 0.189808296 0.193042207
#>  [205] 0.497544960 0.273659283 0.084904282 0.085217228 0.225614578 0.190373457
#>  [211] 0.204829568 0.160693987 0.171075908 0.382622590 0.257797425 0.137757781
#>  [217] 0.078049295 0.028683461 0.090105801 0.095594992 0.372767685 0.373096821
#>  [223] 0.087649455 0.062532400 0.197485454 0.391188199 0.325349085 0.357348619
#>  [229] 0.244150172 0.416951188 0.110066327 0.290457018 0.235108437 0.107407421
#>  [235] 0.207211995 0.297606785 0.464596381 0.196647159 0.505980716 0.652574069
#>  [241] 0.183696910 0.430584968 0.043875933 0.172988055 0.266277774 0.173286774
#>  [247] 0.098567288 0.343552564 0.096246573 0.048039588 0.664175593 0.171536810
#>  [253] 0.426935201 0.294856055 0.529689884 0.113196443 0.215835139 0.376967765
#>  [259] 0.716330568 0.604971404 0.341375097 0.549142648 0.187095785 0.031014460
#>  [265] 0.053832659 0.262106088 0.100526331 0.199757356 0.234322049 0.538748567
#>  [271] 0.790059966 0.487066341 0.321311927 0.658631281 0.759433382 0.039746183
#>  [277] 0.083867670 0.091042252 0.104601696 0.098756019 0.272255707 0.281401642
#>  [283] 0.347044469 0.162117106 0.302254787 0.560852583 0.644834177 0.422632661
#>  [289] 0.088786545 0.118922566 0.067441450 0.100247975 0.053392054 0.016688174
#>  [295] 0.053464734 0.049872286 0.080866272 0.096370485 0.109414109 0.576225237
#>  [301] 0.819528785 0.537929568 0.711585099 0.547058914 0.340763266 0.547107582
#>  [307] 0.331776526 0.030763253 0.078360699 0.063115810 0.213235438 0.243044305
#>  [313] 0.161426020 0.033358132 0.221329696 0.103726958 0.241916642 0.040615818
#>  [319] 0.061335567 0.050562882 0.111108319 0.091582117 0.152152114 0.114117908
#>  [325] 0.120928333 0.196038102 0.244833784 0.175803602 0.048854286 0.178059052
#>  [331] 0.102665103 0.130566585 0.143802836 0.114142418 0.183210525 0.203542421
#>  [337] 0.066762671 0.267469925 0.229010001 0.094571022 0.254180920 0.112030018
#>  [343] 0.142591325 0.171897922 0.104301884 0.130089796 0.254287208 0.135677287
#>  [349] 0.144852185 0.141801937 0.303369121 0.122542818 0.235288500 0.225763872
#>  [355] 0.184686970 0.153574827 0.190386120 0.130098294 0.096057755 0.086968854
#>  [361] 0.000000000 0.200022140 0.191057286 0.137296761 0.141183929 0.128389558
#>  [367] 0.199654874 0.116174272 0.065585390 0.175043788 0.056663937 0.122916171
#>  [373] 0.065997041 0.165359377 0.201411581 0.042023858 0.272654995 0.332395371
#>  [379] 0.334252393 0.137055481 0.280504527 0.154707855 0.193038103 0.164009966
#>  [385] 0.119335029 0.274127589 0.247476379 0.550413517 0.155230344 0.155478978
#>  [391] 0.128274353 0.083143260 0.382893740 0.402915386 0.404660873 0.661962449
#>  [397] 0.224330616 0.255209675 0.118320735 0.103058496 0.188343360 0.167878912
#>  [403] 0.426873243 0.302813180 0.361582623 0.310162996 0.512702889 0.010487472
#>  [409] 0.224098015 0.255916814 0.260483640 0.187980618 0.156260828 0.133847238
#>  [415] 0.115508409 0.137386305 0.180945511 0.252437017 0.047173685 0.162177535
#>  [421] 0.236493621 0.057305697 0.068378039 0.246787140 0.327152623 0.111522902
#>  [427] 0.013089661 0.115842406 0.104878880 0.075174457 0.129599935 0.092261167
#>  [433] 0.086815134 0.108618043 0.103155007 0.037243971 0.094054225 0.310013027
#>  [439] 0.129240973 0.541080797 0.433322261 0.257893245 0.692119172 0.094247342
#>  [445] 0.061757241 0.150515207 0.116046105 0.126588382 0.172659619 0.118425565
#>  [451] 0.177776558 0.255168969 0.243932861 0.041436739 0.124142229 0.116505781
#>  [457] 0.262823088 0.239970445 0.210052404 0.249915712 0.345031254 0.069027883
#>  [463] 0.126275683 0.137857859 0.259540578 0.123832823 0.113249866 0.111959824
#>  [469] 0.187232563 0.346309290 0.113290648 0.131840366 0.097170955 0.346042506
#>  [475] 0.434333568 0.291830898 0.315631399 0.396546859 0.356173066 0.260027203
#>  [481] 0.267051175 0.363359915 0.259244111 0.228133355 0.477881353 0.467390734
#>  [487] 0.530678940 0.371759231 0.135817217 0.228550715 0.198475906 0.293089139
#>  [493] 0.256889972 0.300484650 0.351168644 0.364006023 0.063792175 0.183657698
#>  [499] 0.072313633 0.155566194 0.141924494 0.246747351 0.172170518 0.154337916
#>  [505] 0.124043980 0.474333230 0.534244107 0.161955301 0.117258516 0.189486655
#>  [511] 0.244186693 0.328840007 0.085756984 0.074900711 0.226048702 0.090480310
#>  [517] 0.101306775 0.067318988 0.266142730 0.274339805 0.011335002 0.239097543
#>  [523] 0.314395767 0.017000903 0.053496842 0.065432604 0.052763302 0.071744780
#>  [529] 0.127841342 0.102033480 0.768588505 0.429918712 0.408957629 0.492289973
#>  [535] 0.666905453 0.335039747 0.784289999 0.597250458 0.320067294 0.110823355
#>  [541] 0.044739936 0.173128904 0.359306924 0.790339241 0.350290453 0.890350678
#>  [547] 0.818330450 0.716735881 0.145433452 1.037086347 0.230782662 0.910400067
#>  [553] 0.902610626 0.025016779 0.089174399 0.088907038 0.135032742 0.087001489
#>  [559] 0.060461888 0.050542933 0.036015729 0.067918227 0.149944576 0.155215541
#>  [565] 0.232884967 0.254095356 0.228411541 0.525044110 0.000000000 0.384349541
#>  [571] 0.217679423 0.317193086 0.165196761 0.136261645 0.357599124 0.155872524
#>  [577] 0.173530436 0.094246977 0.193010574 0.431083982 0.088914612 0.135819949
#>  [583] 0.148124773 0.315609862 0.340072297 0.129303596 0.176115867 0.392145355
#>  [589] 0.196201157 0.461561138 0.147057943 0.185007728 1.065647158 0.361065261
#>  [595] 0.477588763 0.434173298 0.386811878 0.451212043 0.469825891 0.374934812
#>  [601] 0.541879082 0.468251902 0.238613119 0.437646417 0.429812547 0.326416479
#>  [607] 0.238012478 0.586793836 0.527150580 0.366164704 0.400755894 0.232504418
#>  [613] 0.481764234 0.287005712 0.455392604 0.369478198 0.410253406 0.514426619
#>  [619] 0.440090597 0.445816403 0.391360788 0.377782440 0.370529142 0.357677258
#>  [625] 0.000000000 0.314942147 0.257556089 0.439502724 0.332744728 0.262001379
#>  [631] 0.319922428 0.403185401 0.358692215 0.318450343 0.211391595 0.270439906
#>  [637] 0.247489239 0.399798000 0.314846076 0.292421734 0.252292095 0.275968279
#>  [643] 0.242584901 0.336224854 0.621504109 0.215238004 0.165776874 0.066397581
#>  [649] 0.013349523 0.019027809 0.074876883 0.400594044 0.389548887 0.294566259
#>  [655] 0.397371966 0.412848498 0.468294775 0.319719883 0.331671019 0.349680602
#>  [661] 0.319932693 0.150961999 0.353477135 0.296320467 0.377308700 0.108333526
#>  [667] 0.323787834 0.192157960 0.188195234 0.173793495 0.306515714 0.752327733
#>  [673] 0.277020580 0.189369453 0.243311419 0.116613875 0.175275705 0.780610703
#>  [679] 0.164625496 1.004630157 0.083303829 0.073240868 0.034630345 0.043017479
#>  [685] 0.021300932 0.046002439 0.052607244 0.056192384 0.098911956 0.013950684
#>  [691] 0.071494030 0.174381746 0.316812153 0.344052906 0.329505288 0.341101050
#>  [697] 0.257800684 0.232485446 0.317057835 0.200563570 0.262495021 0.331010804
#>  [703] 0.316282804 0.294583903 0.209342854 0.141470828 0.120665344 0.147464439
#>  [709] 0.144959042 0.075330769 0.125571973 0.062105683 0.234672819 0.162599701
#>  [715] 0.139207829 0.076282290 0.423278241 0.068085755 0.124563873 0.117235828
#>  [721] 0.135491556 0.198399611 0.169780527 0.124828878 0.094506598 0.180607344
#>  [727] 0.277237142 0.270243074 0.234573412 0.102945089 0.051850969 0.447694350
#>  [733] 0.032520995 0.179412073 0.115286884 0.112097864 0.022942050 0.118590308
#>  [739] 0.228005571 0.119202098 0.258303819 0.221451831 0.116280977 0.205100156
#>  [745] 0.204594020 0.110535477 0.228729511 0.260493695 0.252638156 0.182538487
#>  [751] 0.131792789 0.083305623 0.051098862 0.156545607 0.152176236 0.153268225
#>  [757] 0.199979167 0.120832375 0.028166248 0.249716109 0.342230177 0.249909915
#>  [763] 0.203151915 0.576960478 0.459827552 1.110383248 1.268846239 0.717654428
#>  [769] 0.602776782 0.623552064 0.041429582 0.071759088 0.074911733 0.078915770
#>  [775] 0.016965540 0.088317803 0.216327473 0.345362542 0.408592684 0.465388232
#>  [781] 0.375445365 0.226947615 0.363143973 0.105559880 0.301468687 0.329707933
#>  [787] 0.429294505 0.451895412 0.495683594 0.314640406 0.405452667 0.563646638
#>  [793] 0.399582027 0.596918587 0.521095415 0.131053324 0.149738842 0.151105665
#>  [799] 0.121599603 0.167344826 0.224732396 0.199108132 0.063151287 0.128681116
#>  [805] 0.088312863 0.148138493 0.121013982 0.084659789 0.039898819 0.371550353
#>  [811] 0.401134412 0.319433632 0.365478749 0.337971214 0.266989764 0.370083859
#>  [817] 0.330625004 0.383134479 0.465295632 0.422296989 0.496333272 0.282163028
#>  [823] 0.313984010 0.015432042 0.032989447 0.044165311 0.371765835 0.297149211
#>  [829] 0.155700312 0.436671235 0.296909109 0.134875234 0.166468760 0.164162231
#>  [835] 0.169249320 0.202390694 0.257473952 0.208153897 0.207136291 0.235333577
#>  [841] 0.238913629 0.278994164 0.307480505 0.257533165 0.207972356 0.200344680
#>  [847] 0.256009348 0.067747121 0.552370603 0.124044244 0.380212040 1.064019838
#>  [853] 0.504670547 0.088373600 0.121132529 0.184631942 0.270429967 0.000000000
#>  [859] 0.504521862 0.948227697 0.083812720 0.000000000 0.046924591 0.445448575
#>  [865] 0.262523123 0.283134700 0.437475347 0.391412417 0.395952328 0.424735355
#>  [871] 0.236683578 0.216516278 0.282302844 0.338444679 0.126390595 0.345755997
#>  [877] 0.560418893 0.592232372 0.360097363 0.111996547 0.507145824 0.159003758
#>  [883] 0.634026827 0.122173994 0.072417083 0.126095568 0.145698527 0.147414660
#>  [889] 0.608539706 0.467041373 0.219573353 0.041556832 0.428460817 0.539495516
#>  [895] 0.116786873 0.110131796 0.042713752 0.136731348 0.166564839 0.212394492
#>  [901] 0.195168139 0.133708508 0.079113246 0.055289029 0.050769032 0.049548601
#>  [907] 0.067669965 0.273163592 0.100343037 0.442540510 0.106575526 0.129984121
#>  [913] 0.229284983 0.189611694 0.157658457 0.285686206 0.148959179 0.107563043
#>  [919] 0.212040501 0.135558035 0.074411050 0.087464402 0.232849516 0.239418371
#>  [925] 0.273450874 0.053606552 0.134257049 0.170473070 0.432434203 0.037277824
#>  [931] 0.110711663 0.229565642 0.063083044 0.188659332 0.156133553 0.109557760
#>  [937] 0.110771203 0.155158769 0.027394900 0.160733189 0.041416418 0.027650995
#>  [943] 0.060959291 0.130673107 0.149454045 0.064010017 0.090293293 0.210229998
#>  [949] 0.191052537 0.244311586 0.191958153 0.217966504 0.192858695 0.145116482
#>  [955] 0.215454021 0.062568838 0.067360471 0.041970922 0.037122626 0.125343157
#>  [961] 0.189896205 0.095407313 0.144659151 0.131458140 0.221769519 0.125064312
#>  [967] 0.283637840 0.322604747 0.176750582 0.213053591 0.226832363 0.267053367
#>  [973] 0.310701287 0.186565959 0.290817895 0.137737662 0.226912505 0.634324596
#>  [979] 0.137559845 0.876190493 0.135221161 0.181197964 0.085179323 0.540426336
#>  [985] 0.790052406 0.827274225 0.214797356 0.167652631 0.487475861 0.321514681
#>  [991] 0.134710745 0.222464680 0.185734033 0.177116740 0.276682107 0.330815394
#>  [997] 0.171720242 0.825440489 0.930236402 0.620379070 0.873745459 0.412027657
#> [1003] 0.222926232 0.257569567 0.401684003 0.411263359 0.282262443 0.241589623
#> [1009] 0.422409084 0.372944078 0.460626503 0.244793769 0.295546609 0.234263616
#> [1015] 0.225748584 0.261368092 0.172111088 0.125557305 0.000000000 0.000000000
#> [1021] 0.331549069 0.078624673 0.099273423 0.108773329 0.077119243 0.038423805
#> [1027] 0.098233051 0.036569116 0.036841206 0.227582666 0.094508398 0.410575849
#> [1033] 0.376794882 0.028185922 0.645357743 0.495690949 0.610083753 0.547573442
#> [1039] 0.183011851 0.862354684 0.309492533 0.070843045 0.212444296 0.520260346
#> [1045] 0.031334080 0.196969300 0.046011392 0.115466793 0.158415886 0.115240246
#> [1051] 0.181323478 0.194311399 0.223751814 0.175806865 0.045897692 0.069533659
#> [1057] 0.082207805 0.125940458 0.066952492 0.128229903 0.168119676 0.079030811
#> [1063] 0.120005978 0.052089350 0.082564110 0.056780537 0.216592681 0.097760418
#> [1069] 0.108565821 0.583500696 0.070674985 0.659329534 0.500234885 0.368923529
#> [1075] 0.743215263 0.034186436 0.059607388 0.639116904 0.435188870 0.290215615
#> [1081] 0.260305259 0.183394569 0.437098392 0.567909236 0.517815156 0.505413537
#> [1087] 0.378310582 0.445709430 0.511974951 0.506593674 0.621187521 0.751021090
#> [1093] 0.449655241 0.558705021 0.603752423 0.092532844 0.239110825 0.387028366
#> [1099] 0.777870599 0.240066381 0.311842999 0.567578237 0.400506954 0.707306116
#> [1105] 0.190098896 0.437556448 0.699394405 0.033895943 0.143244976 0.215859222
#> [1111] 0.346908835 0.432364381 0.449664218 0.866903003 0.665384098 0.858802057
#> [1117] 0.649284569 0.821992625 0.691149044 0.502316004 0.712303953 0.463880425
#> [1123] 0.255852759 0.478003502 0.770993380 0.477223116 0.587473101 0.654858643
#> [1129] 0.681162701 0.877326042 0.464860499 0.518027697 0.217343511 0.438520795
#> [1135] 0.283678952 0.437782035 0.352218995 0.622263427 0.623938341 0.415404256
#> [1141] 0.577812606 0.062945419 0.038573121 0.418583360 0.147375842 0.769165615
#> [1147] 0.428425049 0.339459267 0.526639750 0.526803017 0.274410311 0.498776865
#> [1153] 0.488816154 0.478860222 0.524269126 0.546318844 0.391368640 0.318855992
#> [1159] 0.384108800 0.707414664 0.726670571 0.428836670 0.394129721 0.116619037
#> [1165] 0.342827925 0.443864263 0.437627552 0.416695616 0.199208297 0.360708350
#> [1171] 0.401667658 0.517812098 0.295415218 0.050890624 0.365759067 0.118588598
#> [1177] 0.145910646 0.213566693 0.047582835 0.217104413 0.412418656 0.269906080
#> [1183] 0.424748497 0.138243296 0.336338422 0.067013716 0.089630087 0.126197096
#> [1189] 0.024218877 0.154608847 0.031369275 0.124242465 0.068922594 0.072721596
#> [1195] 0.216209136 0.057297629 0.077640760 0.037704471 0.032815852 0.039747199
#> [1201] 0.107346366 0.025314704 0.037125866 0.062190470 0.091379612 0.164116059
#> [1207] 0.136293704 0.305922217 0.158807153 0.010193354 0.025280714 0.051602917
#> [1213] 0.092305500 0.015311925 0.041860705 0.213055268 0.131013033 0.027352922
#> [1219] 0.016866586 0.111735787 0.066944921 0.019340469 0.016928365 0.081246555
#> [1225] 0.037244843 0.085713619 0.215802617 0.178016454 0.102391883 0.253439921
#> [1231] 0.264871851 0.376916832 0.391568374 0.282543563 0.191351206 0.273529708
#> [1237] 0.276524197 0.236915014 0.242161140 0.382619619 0.159040288 0.386229641
#> [1243] 0.284619682 0.458008114 0.490422021 0.319412449 0.301672995 0.697583263
#> [1249] 0.710162794 0.631036514 0.620345918 0.489684593 0.590987397 0.588301887
#> [1255] 0.460461724 0.283806248 0.342589808 0.512540263 0.428564450 0.131651941
#> [1261] 0.255772009 0.260187081 0.216078752 0.242207026 0.185737545 0.154711492
#> [1267] 0.211206882 0.132374071 0.226065036 0.282866723 0.167149069 0.327442241
#> [1273] 0.129468404 0.064523380 0.062738788 0.209350037 0.136641397 0.146823804
#> [1279] 0.213741551 0.180672591 0.225149732 0.073979515 0.066426217 0.184090596
#> [1285] 0.023857783 0.094636810 0.031180380 0.040131363 0.079203023 0.024189187
#> [1291] 0.029753541 0.079539240 0.204087690 0.081030640 0.086832286 0.083461239
#> [1297] 0.065112656 0.081835033 0.113285584 0.252319282 0.144732588 0.233376043
#> [1303] 0.211584520 0.220196247 0.244178571 0.258671554 0.200899347 0.282798502
#> [1309] 0.104002253 0.097086860 0.226416430 0.180327036 0.175221955 0.138339796
#> [1315] 0.130060833 0.234401610 0.252317846 0.189545827 0.206899945 0.246879766
#> [1321] 0.176107042 0.070438526 0.174296613 0.267428381 0.254261610 0.243930024
#> [1327] 0.187449613 0.190515073 0.373210268 0.152151666 0.148947000 0.232848397
#> [1333] 0.248358833 0.082009276 0.218932875 0.177989134 0.155741022 0.206719824
#> [1339] 0.054647379 0.120641766 0.029284396 0.248419206 0.054427927 0.130290441
#> [1345] 0.041741259 0.109306402 0.160237936 0.014868081 0.028163752 0.061443963
#> [1351] 0.023309428 0.044020083 0.044653896 0.057684670 0.053460240 0.031286716
#> [1357] 0.024161213 0.033248651 0.039798277 0.038126336 0.018330421 0.159688256
#> [1363] 0.271229707 0.446669545 0.056088773 0.049751804 0.031276532 0.056800263
#> [1369] 0.136026027 0.158584685 0.078775237 0.149325460 0.069864844 0.111808006
#> [1375] 0.181780660 0.109542101 0.132250110 0.306251867 0.223029738 0.218223715
#> [1381] 0.183701123 0.081056369 0.368489149 0.578055024 0.369804638 0.292424514
#> [1387] 0.161089406 0.351353848 0.930973436 0.860205408 0.076732836 0.190737583
#> [1393] 0.146379458 0.124389402 0.407431432 0.255384857 0.092155115 0.109843839
#> [1399] 0.106542347 0.073346650 0.134705574 0.289417331 0.168722546 0.186858276
#> [1405] 0.381296445 0.264109692 0.399815949 0.219722339 0.262192638 0.267781160
#> [1411] 0.117226152 0.072209127 0.284804463 0.228769851 0.191209242 0.316612756
#> [1417] 0.179138614 0.188513313 0.361271392 0.351815535 0.383140793 0.591576788
#> [1423] 0.272765473 0.147366521 0.227831975 0.066069811 0.113506726 0.110761741
#> [1429] 0.141330704 0.128188924 0.041762618 0.023964307 0.030514972 0.087105540
#> [1435] 0.086126621 0.066275620 0.064122074 0.079872042 0.098404242 0.080568272
#> [1441] 0.128482707 0.115000223 0.257700445 0.175140893 0.197205751 0.290580726
#> [1447] 0.267179693 0.267136904 0.253027969 0.317481499 0.276892629 0.161662583
#> [1453] 0.379350377 0.096850595 0.180699756 0.147861206 0.124199741 0.090147135
#> [1459] 0.144665618 0.150877588 0.127526886 0.268272875 0.083902790 0.137473039
#> [1465] 0.175449128 0.118831157 0.233358140 0.138801971 0.097220446 0.050649918
#> [1471] 0.050999638 0.696908447 0.197578512 0.121046312 0.125949885 0.134747447
#> [1477] 0.096907647 0.333902407 0.181488670 0.132297923 0.156097392 0.128574180
#> [1483] 0.307023567 0.277345035 0.171101886 0.188460046 0.189500650 0.559292685
#> [1489] 0.350218753 0.317731235 0.346339702 0.346292944 0.437271917 0.301699271
#> [1495] 0.471027146 0.293302289 0.354791374 0.168753771 0.235273344 0.218124438
#> [1501] 0.181385006 0.182278426 0.202932563 0.136918032 0.257556732 0.251331252
#> [1507] 0.421484498 0.329853301 0.242883675 0.414578818 0.351808189 0.353076966
#> [1513] 0.258533535 0.248941290 0.223820218 0.189774138 0.190053680 0.202946687
#> [1519] 0.218768784 0.240831578 0.257366199 0.179280832 0.177862709 0.237476699
#> [1525] 0.247193618 0.292463584 0.487633035 0.343130626 0.519108861 0.587582943
#> [1531] 0.236243861 0.326641355 0.223357474 0.189848244 0.158277251 0.215822915
#> [1537] 0.212279039 0.341388798 0.462601481 0.333746734 0.223696310 0.255631644
#> [1543] 0.373100963 0.269586788 0.208119703 0.180311408 0.189759115 0.241549822
#> [1549] 0.166399549 0.134680809 0.166119064 0.193479846 0.326470710 0.212040979
#> [1555] 0.180344497 0.176497726 0.474653441 0.360171936 0.478470444 0.202613115
#> [1561] 0.116500973 0.382801875 0.419882478 0.409595817 0.390240058 0.411704863
#> [1567] 0.374936332 0.453149296 0.379827143 0.389401964 0.412497332 0.227474010
#> [1573] 0.234613371 0.232061640 0.314124277 0.307633697 0.232715190 0.265033663
#> [1579] 0.382716736 0.365607228 0.364394986 0.327363290 0.328963614 0.312428648
#> [1585] 0.424403248 0.447744365 0.377996888 0.000000000

# Using dplyr
require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

mutate(bay_race, divergence_score = divergence(white, hispanic,
  asian, black, all_other))

#> Warning: Population parameter not set; assuming equal populations.

#> # A tibble: 1,588 x 9
#>    fips     total_pop hispanic white  black  asian all_other county             
#>    <chr>        <dbl>    <dbl> <dbl>  <dbl>  <dbl>     <dbl> <chr>              
#>  1 0600140…      2937   0.0398 0.708 0.0477 0.155     0.0497 Alameda County, Ca…
#>  2 0600140…      1974   0.0765 0.783 0.0157 0.0740    0.0507 Alameda County, Ca…
#>  3 0600140…      4865   0.0820 0.669 0.105  0.0861    0.0573 Alameda County, Ca…
#>  4 0600140…      3703   0.0897 0.655 0.121  0.0729    0.0618 Alameda County, Ca…
#>  5 0600140…      3517   0.0967 0.506 0.265  0.0591    0.0734 Alameda County, Ca…
#>  6 0600140…      1571   0.0802 0.427 0.391  0.0509    0.0503 Alameda County, Ca…
#>  7 0600140…      4206   0.104  0.295 0.492  0.0592    0.0502 Alameda County, Ca…
#>  8 0600140…      3594   0.102  0.316 0.407  0.119     0.0556 Alameda County, Ca…
#>  9 0600140…      2302   0.133  0.310 0.437  0.0556    0.0647 Alameda County, Ca…
#> 10 0600140…      5678   0.142  0.223 0.502  0.0736    0.0594 Alameda County, Ca…
#> # … with 1,578 more rows, and 1 more variable: divergence_score <dbl>

# divergence alsow works with percentages as long as you have
# population totals by observation
bay_race %>%
  mutate_at(vars(hispanic:all_other), list(~(./total_pop))) %>%
  mutate(divergence_score = divergence(white, hispanic, asian,
  black, all_other, weights = total_pop))

#> Warning: parameter `weights` is deprecated, use `population`

#> # A tibble: 1,588 x 9
#>    fips    total_pop  hispanic    white    black   asian all_other county       
#>    <chr>       <dbl>     <dbl>    <dbl>    <dbl>   <dbl>     <dbl> <chr>        
#>  1 060014…      2937 0.0000136  2.41e-4  1.62e-5 5.29e-5 0.0000169 Alameda Coun…
#>  2 060014…      1974 0.0000388  3.97e-4  7.96e-6 3.75e-5 0.0000257 Alameda Coun…
#>  3 060014…      4865 0.0000169  1.38e-4  2.16e-5 1.77e-5 0.0000118 Alameda Coun…
#>  4 060014…      3703 0.0000242  1.77e-4  3.27e-5 1.97e-5 0.0000167 Alameda Coun…
#>  5 060014…      3517 0.0000275  1.44e-4  7.54e-5 1.68e-5 0.0000209 Alameda Coun…
#>  6 060014…      1571 0.0000511  2.72e-4  2.49e-4 3.24e-5 0.0000320 Alameda Coun…
#>  7 060014…      4206 0.0000248  7.00e-5  1.17e-4 1.41e-5 0.0000119 Alameda Coun…
#>  8 060014…      3594 0.0000284  8.79e-5  1.13e-4 3.32e-5 0.0000155 Alameda Coun…
#>  9 060014…      2302 0.0000579  1.35e-4  1.90e-4 2.42e-5 0.0000281 Alameda Coun…
#> 10 060014…      5678 0.0000250  3.94e-5  8.83e-5 1.30e-5 0.0000105 Alameda Coun…
#> # … with 1,578 more rows, and 1 more variable: divergence_score <dbl>