Theil's Index of Entropy

Entropy is used to measure the the extent to which multiple distributions conform to a baseline.

entropy(
  ...,
  weights = "sum",
  sumPercent = "weights",
  entropy_type = "entropy",
  scale = FALSE,
  summed = FALSE,
  na.rm = TRUE
)

Arguments

...
Population vectors for every group included in the divergence calculation.

weights
Either a vector of weights summing to one, or the total population for each observation in .... Can be any of:

A numeric vector

The length of each vector provided in each entry in ... (i.e. the column-wise length of ...). Can be either population or weights by observation

sum

Sets weights to the rowwise sum of ...

none

Weighs each observation evenly (1/length). Note that if ... is a set of percentages summing to 1 for each row, sum and none are equivalent.

sumPercent
The percentage of each group in the larger population (i.e. the population composed of the sum of groups provided in ...). Can be any of:

A numeric vector

Each entry represents the total population of the with one entry for each group/vector provided in .... (i.e. the rowwise length of ...). Note that for this to work correctly, each group must be provided in the same order in sumPercent as in ... Useful if the sum of sumPercent is less than 100%, as divergence will otherwise force

weights

Default. Uses the value of weights to construct total population proportions. If weights is set to none, the total population proportions will be the unweighted average of the percentages in each observation.

entropy_type
One of:

"entropy"

Default. \(E_{i} = \Sigma (X_{im} * ln(1/X_{im})\) where Xim is the proportion of racial group within the geography i.

"information_theory"

Theil's information theory index

scale
Scale entropy scores from 0-1. Setting scale to TRUE ignores the entropy_type (set to "entropy") and summed parameters.

summed
If TRUE, will return a single summary statistic. (Or one value per group if specifying dplyr::group_by.) If FALSE (default), will return a vector equaling the length of the input vectors. If 'weighted' (only for divergence and information theory), returns a vector as in FALSE, but with pre-weighted values, such that sum(divergence(..., summed = 'weighted)) is equivalent to divergence(..., summed = T).

na.rm
logical. Should missing values (including NaN) be removed? Used only if summed is set to TRUE.

Source

Theil, Henri. 1972. Statistical Decomposition Analysis.

Value

A single value if summed==TRUE, or a vector equaling the length of the inputs. Note that if entropy_type == "index", and summed is FALSE, then the returned vector will be entropy index, unweighted by population

See also

Examples

library(rsegregation)
entropy(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, bay_race$all_other, weights = bay_race$total_pop)

#>    [1] 0.9566644 0.7969746 1.0859266 1.1121719 1.2816122 1.2348325 1.2624260
#>    [8] 1.3773304 1.3313959 1.3177736 1.3787653 1.2199007 1.4615506 1.3515930
#>   [15] 1.2757467 1.3737796 1.4518171 1.1817099 1.4368619 1.2489296 1.0412827
#>   [22] 1.2220339 1.2663997 1.4051752 1.3755270 0.5085328 1.4338626 1.2123693
#>   [29] 1.4394896 1.4778355 1.3825896 1.3833122 1.3414567 1.2515983 1.1574508
#>   [36] 1.2782505 1.2571103 1.0640105 1.0571691 0.9395430 0.9172207 1.0402935
#>   [43] 0.9556011 0.9725084 0.9464187 0.9875859 1.4041509 1.2810355 1.1066259
#>   [50] 1.0516796 1.4175284 1.4417597 1.4531823 1.3884419 1.3269170 1.3721921
#>   [57] 1.4475982 1.4049155 1.2487191 1.2362252 1.2271264 1.3469760 1.2170678
#>   [64] 1.1886146 1.0706730 1.3322115 1.4884093 1.3607427 1.4573588 1.4601957
#>   [71] 1.3383743 1.4113513 1.4845217 1.4614611 1.2212170 1.3054676 0.8800993
#>   [78] 1.1231843 1.0284870 1.1240041 1.4177022 1.3031458 1.4433086 1.4676021
#>   [85] 1.3255902 1.4119622 1.2598101 1.3812554 1.0687016 1.0222064 1.0516074
#>   [92] 1.0952942 1.1069063 0.9763299 0.9963739 0.9712796 1.0582725 0.9714586
#>   [99] 0.9286148 0.8881621 0.9663981 1.0975260 1.0784676 1.2283630 1.2485657
#>  [106] 1.1497812 1.0957255 0.9897953 1.1849689 1.1503189 1.1018838 1.2697794
#>  [113] 1.2413063 1.2773615 1.2290170 0.9880787 0.8273490 0.7648703 0.7984622
#>  [120] 0.6902946 0.7537593 0.8183888 1.0162652 0.8517331 1.1534537 1.3817493
#>  [127] 1.4548175 1.2889772 1.1575211 1.1947036 1.1008021 1.1824004 1.1823207
#>  [134] 1.2939223 1.3641844 1.1930503 1.3789880 1.4326986 1.4100849 1.4348640
#>  [141] 1.3288734 0.9970738 1.1917571 1.0143403 0.6497142 1.2630261 0.9547562
#>  [148] 1.4357097 1.3830825 1.0920561 1.3603973 1.4180034 1.4967398 0.8508674
#>  [155] 0.9257936 1.0288332 1.3582058 1.3526813 1.4648551 1.3292936 1.2638929
#>  [162] 1.2233992 1.4200580 1.1874206 1.2084739 1.2337526 1.1169295 1.3249818
#>  [169] 1.3204328 1.3059091 1.4736282 1.2063413 1.0399133 1.0861335 1.0611130
#>  [176] 1.0666057 1.4729361 1.2401820 1.1942204 1.1975276 1.4033114 1.3350036
#>  [183] 1.4236106 1.3273584 1.3881453 1.4081306 1.4240358 1.4112350 1.3459262
#>  [190] 1.3919113 1.4916431 1.3506054 1.4171100 1.4038352 1.3947975 1.4380746
#>  [197] 1.4713506 1.4687615 1.3307735 1.2868088 1.3234832 1.3996490 1.3929309
#>  [204] 1.4490949 1.3238029 1.3962044 1.4898195 1.3551925 1.5010161 1.5210501
#>  [211] 1.4794614 1.4923480 1.4035755 1.2324828 1.2702046 1.3359120 1.3440899
#>  [218] 1.3369240 1.2620869 1.3620630 1.2859920 1.2632731 1.4570991 1.3260086
#>  [225] 1.5021489 1.2558171 1.4333863 1.2236057 1.4243024 1.1506978 1.5076399
#>  [232] 1.3203798 1.4090460 1.5001864 1.5218358 1.2400294 1.2296945 1.4214262
#>  [239] 1.2923816 1.0275971 1.4512196 1.2145985 1.4274266 1.3837867 1.3153907
#>  [246] 1.4049547 1.4817767 1.3591264 1.4076516 1.4138891 0.8722777 1.3716245
#>  [253] 1.1772057 1.3237352 1.1350424 1.3899775 1.3382275 1.2686403 0.8618577
#>  [260] 0.9881536 1.2777730 1.0412336 1.3902297 1.2690377 1.2846785 1.2195250
#>  [267] 1.3312135 1.2938012 1.2591091 1.0619007 0.7309819 1.0475133 1.2161443
#>  [274] 0.8869281 0.7640380 1.3674627 1.4030450 1.3888404 1.2948235 1.3428446
#>  [281] 1.2510764 1.1999383 1.1760795 1.3689209 1.2216673 0.8752574 0.7677809
#>  [288] 0.9672229 1.3194096 1.4328444 1.3351903 1.3289544 1.2376450 1.3805945
#>  [295] 1.2068619 1.3242970 1.3329127 1.2997845 1.2998998 0.8444220 0.6450404
#>  [302] 0.8984425 0.7362143 0.8926510 1.0433112 0.9442975 1.0806453 1.3454085
#>  [309] 1.3777199 1.4009327 1.3219501 1.2310269 1.3618146 1.3688851 1.2959541
#>  [316] 1.3194221 1.5045205 1.2335007 1.1855500 1.1967328 1.0834087 1.1347248
#>  [323] 0.9982352 1.0813487 1.0428359 0.8887186 0.8039439 0.9607194 1.1722943
#>  [330] 0.9470695 1.0574230 1.0111302 1.2643644 1.0805888 1.1052509 0.8837970
#>  [337] 1.3641395 1.1333636 1.1804687 1.0895459 0.7996069 1.0648617 0.9972814
#>  [344] 0.9664486 1.0702422 1.0754134 1.0710518 0.9987021 1.0093875 1.0109904
#>  [351] 0.7350754 1.1578928 0.8402555 0.8556889 0.9652592 0.9709924 0.9192574
#>  [358] 1.1561357 1.0661798 1.2910921 0.0000000 0.9460429 1.1583275 1.1775650
#>  [365] 1.2235795 1.2939797 1.5376302 1.1336831 1.3398180 1.1689947 1.4356921
#>  [372] 1.1396284 1.2414045 1.0530894 0.9226577 1.2637111 0.9830448 0.7065768
#>  [379] 0.7098138 1.0949971 1.2328311 1.1902880 1.2007477 1.2723173 1.2103100
#>  [386] 1.1658724 1.2708318 1.2940121 1.2639933 1.3879849 1.2457679 1.4687203
#>  [393] 1.4252009 1.2866885 1.3058130 1.3102872 1.4411993 1.4528363 1.4812310
#>  [400] 1.4165588 1.3467945 1.3359151 1.2577552 1.2753051 1.2774689 1.2956912
#>  [407] 0.9427987 1.3826993 1.2322788 0.8489201 0.8955667 0.9667865 1.1219222
#>  [414] 1.0678571 1.0698969 1.0404081 0.9304500 0.8223569 1.2296017 0.9521027
#>  [421] 0.8468091 1.1970690 1.1371901 0.8393112 0.6955858 1.1627920 1.3169986
#>  [428] 1.1710649 1.1374522 1.1469170 1.0515543 1.1127324 1.1002106 1.0618976
#>  [435] 1.1184322 1.2448109 1.1281695 0.7257662 1.0645012 1.0004974 1.0806242
#>  [442] 1.1283245 0.8208169 1.1089281 1.1926308 0.9879933 1.2907760 1.0505631
#>  [449] 0.9514912 1.0693053 0.9368530 0.7944265 0.8395286 1.2105247 1.0254691
#>  [456] 1.0418792 0.7968869 0.8436603 0.8786751 0.8156207 0.6651655 1.1405213
#>  [463] 1.0115823 1.0052912 0.7997727 1.0647572 1.1544797 1.1822551 0.9370836
#>  [470] 0.6533923 1.0975385 1.0426906 1.1156415 0.6670703 0.5267302 0.7323329
#>  [477] 0.7000382 0.5888844 0.6507321 0.8210851 0.7934270 0.6503686 0.8022719
#>  [484] 0.8702197 0.4513784 0.4750674 0.3820325 0.6352001 1.0480054 0.8589241
#>  [491] 0.9121319 0.7705905 0.8168180 0.7483211 0.7046703 0.6524234 1.3371050
#>  [498] 1.5438586 1.4473103 1.5098882 1.4993862 0.8189329 0.9882090 1.1294954
#>  [505] 1.3806953 0.9758217 0.9262248 1.4350638 1.0519883 0.9373557 0.8348130
#>  [512] 0.6980153 1.4985893 1.4328406 0.9238515 1.4452912 1.5137661 1.4780899
#>  [519] 1.4866786 1.3791736 1.3568342 1.4623443 1.3454913 1.4141416 1.4489077
#>  [526] 1.4679094 1.4914757 1.3156832 1.5027009 1.4159234 1.1337524 1.4810816
#>  [533] 1.2405745 1.1638911 1.3467240 1.3397180 0.7969545 0.9964187 1.4281067
#>  [540] 1.4967269 1.4208047 1.4149602 1.2269651 0.8288258 1.3030446 0.7557968
#>  [547] 1.0774617 1.0425883 1.0416099 0.9670939 1.4976347 1.1380394 1.2430357
#>  [554] 1.4201219 1.2997248 1.1859329 1.2260970 1.4725943 1.3275933 1.3138906
#>  [561] 1.2795696 1.4303078 1.0638965 1.0222513 0.8778399 0.8753009 1.4445394
#>  [568] 1.2389923 0.0000000 0.6027108 0.8595073 0.7100527 0.9560349 1.0800752
#>  [575] 0.6464675 1.0047049 0.9569870 1.1951052 0.9387663 0.5348577 1.1491449
#>  [582] 1.0198495 0.9959712 0.7136558 0.6908610 1.0098832 1.0164713 0.5920685
#>  [589] 0.9042162 0.5053581 1.0335738 0.9634420 0.4243783 1.0341900 0.5056130
#>  [596] 0.5607278 0.6221117 0.5162042 0.5152427 0.6387010 0.3978748 0.4948529
#>  [603] 0.8552082 0.5501581 0.5520057 0.7139350 0.8608542 1.2746832 0.4105161
#>  [610] 0.6537818 0.5990282 0.8623144 0.4885573 0.7624847 0.5215695 0.6551241
#>  [617] 0.5813172 1.4281418 0.5231957 0.5255556 0.6111890 0.6526167 0.6326766
#>  [624] 0.6554360 0.0000000 0.8601989 0.9273160 0.8383725 0.8926278 0.9020540
#>  [631] 0.7201309 0.8611102 0.8573824 0.9051410 1.0212719 0.8721676 0.8615948
#>  [638] 0.5756897 0.8579399 0.8791390 0.8387533 0.7994106 0.9015129 0.6919273
#>  [645] 0.7305965 1.2312252 1.3688728 1.4410485 1.3928005 1.3483426 1.4386317
#>  [652] 0.5792632 0.6093671 0.7757447 0.6068429 0.5544515 0.4830695 0.7054642
#>  [659] 0.7332687 0.8600058 0.7194251 0.9921449 0.6809013 0.7450960 0.8327747
#>  [666] 1.2240443 0.6932427 0.9602709 0.9519950 0.9677951 1.0279785 0.6933701
#>  [673] 0.9629365 0.9226183 1.0309849 1.1502862 1.0199167 0.6498984 1.2320050
#>  [680] 0.4762133 1.1669810 1.1570098 1.2945585 1.2726149 1.3767000 1.4664264
#>  [687] 1.4353569 1.3321718 1.4200745 1.3890909 1.4942639 1.4315261 0.7060187
#>  [694] 0.6503388 0.6852650 0.6652908 0.7853142 0.8421464 0.6913540 0.8968130
#>  [701] 0.8079441 0.6903003 0.7060579 0.7548187 0.8813082 1.0541178 1.0996797
#>  [708] 1.0355555 1.0209327 1.3128411 1.1306629 1.2317610 1.4208192 1.1272199
#>  [715] 1.3506769 1.2216627 1.4180201 1.1796745 1.3395084 1.1637065 1.0867349
#>  [722] 0.9415376 1.0098647 1.1302550 1.2297034 0.9623916 0.7659642 0.8148646
#>  [729] 0.8764336 1.4332557 1.2899075 0.9663329 1.3504707 1.5476404 1.4221968
#>  [736] 1.3214172 1.2755627 1.0327262 0.8505523 1.0445754 0.8170163 0.8670898
#>  [743] 1.0407995 1.1571748 1.1503974 1.0585928 0.8644006 0.8172524 0.8332772
#>  [750] 0.9265957 1.0045504 1.1600363 1.2710956 0.9717399 1.0099711 0.9829484
#>  [757] 0.9175532 1.2134717 1.3429868 1.0881599 1.0993680 1.1063208 1.1662517
#>  [764] 1.0910074 1.3331882 1.1574842 1.1908113 1.2778392 1.0569352 1.4023306
#>  [771] 1.2267325 1.1887849 1.1651300 1.2774739 1.3751822 1.3844550 1.2004360
#>  [778] 1.1527493 1.0970880 1.0242460 1.1977778 1.3347343 1.1229628 1.3130640
#>  [785] 1.1872442 1.1358363 1.0580338 1.0569583 0.9834322 1.1921082 1.0774451
#>  [792] 1.0794092 1.2204330 0.9613428 1.2955512 1.0546007 1.0278401 1.0807309
#>  [799] 1.1175062 1.1188644 1.0660378 1.0616467 1.2887980 1.0594844 1.1506550
#>  [806] 1.0685439 1.1437464 1.1935879 1.2968720 1.1863976 1.2006420 1.3183085
#>  [813] 1.2908155 1.3334052 1.0845459 0.9877359 1.0161131 0.9769454 0.8977923
#>  [820] 0.9442554 0.8842000 1.0372416 0.9980518 1.3721169 1.3392309 1.3646427
#>  [827] 0.9858741 1.0904576 1.1415464 0.9453652 1.0730725 1.0938852 1.0934914
#>  [834] 1.1155116 1.0727365 1.0855914 0.8395455 1.1256407 1.1041048 1.0881505
#>  [841] 1.1099053 1.0534602 1.0307672 1.0888735 1.1072593 1.0980757 0.8608294
#>  [848] 1.1963946 1.4575012 1.1929627 1.2808241 0.4549140 1.3679110 1.4107672
#>  [855] 1.1306154 1.0623242 0.8640216 0.0000000 1.2153746 1.2219670 1.2734582
#>  [862] 0.0000000 1.2596428 1.1853699 1.2114344 1.2681837 1.0898376 1.1279556
#>  [869] 1.1317335 1.1151150 1.3166145 1.1056909 1.2405450 1.0976825 1.3144685
#>  [876] 1.1681171 0.9483154 0.9116090 1.1937796 1.3149618 1.0193791 1.2926902
#>  [883] 0.9089980 1.3079806 1.3181605 1.3592425 1.3361378 1.2903062 0.9220704
#>  [890] 1.0395783 1.2790222 1.3057029 1.0925457 1.0202411 1.3745655 1.3552761
#>  [897] 1.2207012 1.0519415 0.9769471 0.9353441 0.9378916 1.0396113 1.3349749
#>  [904] 1.2816201 1.2052511 1.2031398 1.1467579 1.2661530 1.3701382 1.1905649
#>  [911] 1.2874214 1.0974117 1.0457963 1.0785137 1.1744712 1.0173728 1.0666242
#>  [918] 1.0863686 0.8698357 0.9958927 1.1290720 1.1019484 0.9000972 0.9080831
#>  [925] 0.7717046 1.2569595 1.2213741 1.3518500 1.1653939 1.2672545 1.0709428
#>  [932] 0.8360079 1.1568185 0.9155378 1.0419271 1.1361664 1.0734131 0.9875788
#>  [939] 1.2710883 0.9706026 1.2056434 1.2922388 1.1972485 1.3855265 1.3749300
#>  [946] 1.3114285 1.1485351 1.0866951 1.0510946 1.0957809 1.1555710 1.1002927
#>  [953] 1.1650666 1.1681020 1.1252478 1.3716164 1.1840397 1.2526417 1.2988140
#>  [960] 1.0461039 0.9585177 1.1198175 1.0017848 1.0431272 0.8739212 1.0400865
#>  [967] 0.7562487 0.7210975 0.9387088 0.8830887 0.8857181 0.7788963 0.7211434
#>  [974] 0.9361922 0.7391217 1.0235840 1.1029281 0.8837727 1.3990346 0.6191271
#>  [981] 1.1854623 1.1180747 1.1508973 0.9043002 0.6799918 0.6579772 1.0688958
#>  [988] 1.2016086 0.9603564 1.0475140 1.1043362 0.8684760 0.9207383 0.9304669
#>  [995] 0.7789790 0.7114300 1.0060321 0.9729747 1.0082349 1.2306609 0.8996991
#> [1002] 1.1527797 0.8744779 0.8055654 0.5810336 0.5962942 0.7802400 0.8529755
#> [1009] 0.5417145 0.6196093 0.4886103 0.9457447 0.7597049 0.8589080 0.9141788
#> [1016] 0.8455509 0.9822400 1.3276442 0.0000000 0.0000000 1.1281525 1.3307695
#> [1023] 1.2428207 1.1831286 1.1375816 1.2619147 1.3256011 1.3586415 1.4429349
#> [1030] 1.2147384 1.3388544 1.0171167 1.0599224 1.2748314 0.8512301 0.9269818
#> [1037] 0.8962201 0.9274763 1.2627322 0.6122952 1.0017865 1.3654270 1.1402050
#> [1044] 0.9340213 1.2331869 1.2375221 1.3766660 1.0694328 0.9801379 1.1121236
#> [1051] 0.9328334 0.9120688 0.8445445 0.9524009 1.2176660 1.1476347 1.1171443
#> [1058] 1.0369685 1.1468225 1.0300525 0.9445939 1.1335937 1.0492071 1.1746934
#> [1065] 1.1653879 1.1984634 0.8510316 1.0838764 1.0655916 0.9348029 1.3815986
#> [1072] 0.8735344 1.0121162 1.1095742 0.7357731 1.3039360 1.3247245 0.8781773
#> [1079] 1.0942564 1.2022411 1.2533375 1.3308549 1.0812195 0.9490967 1.0027878
#> [1086] 1.0225768 1.1198318 1.0715570 0.9919274 1.0009862 0.9190216 0.7565833
#> [1093] 1.0820186 0.9729590 0.9196125 1.3422581 1.2157556 1.1528902 0.7499554
#> [1100] 1.2546925 1.1958689 0.9606105 1.1058494 0.8002667 1.2616705 1.0479012
#> [1107] 0.8164865 1.3050159 0.9926863 1.2647054 1.0616932 1.0982222 1.0870004
#> [1114] 0.6630121 0.8922474 0.6733232 0.8783828 0.7552122 0.8170938 1.0174746
#> [1121] 0.7742393 1.0482698 1.1296829 1.0283382 0.7543931 1.0515615 0.9410516
#> [1128] 0.8749831 0.8497604 0.6431376 1.0702356 1.0114632 1.2526224 1.0654911
#> [1135] 1.2048726 1.0108888 1.1185526 0.8864059 0.9200025 1.0656915 0.9034322
#> [1142] 1.2225722 1.2985805 1.0701448 1.2508596 0.7438884 1.0664530 1.1458170
#> [1149] 0.9947360 0.9779513 1.1795262 1.0176979 0.9819056 1.0170036 0.9996304
#> [1156] 0.9602107 1.1076767 1.1303696 1.1411519 0.7722302 0.7713114 1.0386294
#> [1163] 1.1021324 1.2191971 1.1802226 1.0069581 1.0641124 1.1071479 1.2019930
#> [1170] 1.1752072 1.1391613 0.9802671 1.1986669 1.2607836 1.0819458 1.3271036
#> [1177] 1.2749596 1.2039863 1.1870912 1.2593767 1.0580769 1.1842623 0.9869901
#> [1184] 1.3264375 1.0618466 1.4027284 1.3283561 1.3219339 1.2660352 1.3108039
#> [1191] 1.2841677 1.3080940 1.2767166 1.2671861 1.1992497 1.2852436 1.3004094
#> [1198] 1.2102184 1.2554872 1.2140138 1.0684221 1.2387003 1.2343377 1.1843097
#> [1205] 1.1842899 1.1483610 1.1362167 1.1582323 1.1748962 1.3661140 1.2444494
#> [1212] 1.3724602 1.3447190 1.3115095 1.2698180 1.2344912 1.0213576 1.2594153
#> [1219] 1.3102614 1.0685275 1.1884954 1.2692130 1.2912491 1.1225319 1.2134136
#> [1226] 1.1155703 0.8785707 0.9248594 1.0642906 0.8127958 0.8021117 0.6104385
#> [1233] 0.6127404 0.7656603 0.9180617 0.7772105 0.7752076 0.9136259 0.8505111
#> [1240] 0.9094882 1.0198960 0.8899580 0.9352005 0.8751820 0.8453703 0.9791359
#> [1247] 1.0762518 0.7408223 0.7241364 0.7788447 0.7719001 0.8595899 0.8285298
#> [1254] 0.7939354 0.9174517 1.0885525 1.0461251 0.8654722 0.9454122 1.1722372
#> [1261] 1.0485135 1.0869308 0.9855437 0.9200533 1.1953143 1.1271342 0.9788180
#> [1268] 1.1462580 1.1303748 1.1525746 1.1566040 1.1282471 1.2754380 1.2326529
#> [1275] 1.2381326 1.2245919 1.2520290 1.2623380 1.2621062 1.2296587 1.1636354
#> [1282] 1.2411900 1.2895344 1.2869474 1.3576671 1.2062827 1.2816869 1.2328597
#> [1289] 1.1365524 1.3016626 1.3077302 1.2345290 1.2058061 1.2797207 1.2660152
#> [1296] 1.1200471 1.1447239 1.1429472 1.0733700 0.8435796 1.0283458 0.9056513
#> [1303] 0.9438706 0.8968686 0.8703118 0.8261611 0.9606343 0.8003652 1.0955419
#> [1310] 1.2139263 0.9982098 0.9881926 1.0694822 1.0697553 1.1590133 0.9208939
#> [1317] 0.8486298 0.9370453 0.9054808 0.8326858 0.9714468 1.3840761 0.9876207
#> [1324] 0.8571574 0.8662560 0.8878638 0.9559395 0.9136890 0.6381037 1.0817718
#> [1331] 1.0698627 1.0223165 0.8022730 1.1254004 1.0294094 0.9372979 0.9925544
#> [1338] 1.1712535 1.1575897 1.2965995 1.2428043 1.2412601 1.3428272 1.3461496
#> [1345] 1.3441603 1.3324175 1.3122474 1.4040128 1.2575937 1.4047771 1.3319816
#> [1352] 1.2813767 1.2064601 1.1983956 1.2939014 1.2536263 1.2893148 1.2685133
#> [1359] 1.3287322 1.2752987 1.3641179 1.3184375 1.2600583 1.0797964 1.1884294
#> [1366] 1.2290287 1.2461381 1.2831593 1.1739900 0.9572059 1.1195640 1.0058113
#> [1373] 1.1454960 1.0776790 1.1859582 1.1674735 1.0799430 1.0390630 1.0803139
#> [1380] 1.0084167 1.0499918 1.1651597 1.0222270 0.8512338 0.9404739 1.0241919
#> [1387] 1.1042085 0.9353074 0.5349061 0.5975292 1.3102662 0.9970542 1.5214327
#> [1394] 1.4096728 1.3214815 1.3650287 1.5011251 1.5277613 1.4183783 1.3923384
#> [1401] 1.4951942 1.4970858 1.4721990 1.4743377 1.4150033 1.4693520 1.4935942
#> [1408] 1.3814465 1.4127372 1.4412047 1.2097861 1.4441230 1.4134756 1.4901554
#> [1415] 1.4908965 1.5275784 1.5082140 1.4340287 1.3306130 1.5107832 1.4455469
#> [1422] 1.4045897 0.8053251 1.0901273 1.4736675 1.2338048 1.1011904 1.1517935
#> [1429] 1.0765560 1.0930103 1.3044837 1.4567905 1.3733991 1.3539984 1.1321284
#> [1436] 1.4441379 1.3856714 1.4295122 1.4788887 1.5129710 1.5518649 1.5298390
#> [1443] 1.3259500 1.4148846 1.3520857 1.3850180 1.4598956 1.3823596 1.4103704
#> [1450] 1.3461502 1.4222673 1.4917609 1.4782088 1.5327362 1.5573581 1.5327183
#> [1457] 1.5220721 1.5407733 1.4484129 1.2583531 1.3074904 0.8240457 1.1619855
#> [1464] 1.1938793 1.0135350 1.2085564 0.9074330 1.1090899 1.2174379 1.3521024
#> [1471] 1.3716299 1.2790588 0.9704350 1.2971238 1.2282766 1.2368298 1.2513831
#> [1478] 0.7302084 1.0271716 1.2021827 1.1673823 1.0785977 0.8909784 1.0135734
#> [1485] 1.0320647 1.0967205 0.9539388 1.1573733 0.6681975 0.7135356 0.6538253
#> [1492] 0.6535533 0.5239592 0.8756409 0.8205066 0.8463193 0.6555089 1.0179489
#> [1499] 0.8483664 0.9445480 0.9128714 0.9397605 0.8955076 1.0765828 0.8109740
#> [1506] 0.8475297 0.5622110 0.6919100 0.9150837 0.5978409 0.6766604 0.6788971
#> [1513] 0.8477207 0.8513986 0.9023965 1.0255514 1.0511761 0.9438016 0.8819657
#> [1520] 0.8431756 0.8226746 0.9682909 0.9444009 1.1244676 1.0892694 0.7751791
#> [1527] 0.4565162 0.7061668 0.4059841 0.3123462 0.8909030 0.7370755 0.9491919
#> [1534] 1.0224373 1.0399533 0.9344157 0.9357214 0.7096729 0.5087840 0.6768861
#> [1541] 0.8933281 0.8484422 0.6206484 0.8237003 0.9355287 0.9876980 1.1593710
#> [1548] 1.1636033 0.9956593 1.0691993 0.9756525 1.1347162 0.9960601 1.0623049
#> [1555] 0.9679951 1.0004593 0.9540755 1.0265165 0.9003013 1.1641552 1.1998787
#> [1562] 0.6254974 0.5762782 0.5909073 0.6140472 0.5757772 0.6337571 0.5554929
#> [1569] 0.6749822 0.6238346 0.5920758 0.9721617 0.8869645 0.9003470 0.7335529
#> [1576] 0.8998174 0.9254522 0.8319649 0.8415846 0.6530425 0.6517500 0.7598193
#> [1583] 0.8823980 0.7621910 0.5683134 0.5220813 0.6987997 0.0000000