Isolation Index

The summation is over all the component geographic parts of the larger geographic entity for which the isolation index is calculated.

isolation(group, population, summed = TRUE, na.rm = TRUE)

Arguments

group
A numeric vector of the proportion of the population

population
A numeric vector the length of group with its proportion of population.

summed
If TRUE, will return a single summary statistic. (Or one value per group if specifying dplyr::group_by.) If FALSE (default), will return a vector equaling the length of the input vectors. If 'weighted' (only for divergence and information theory), returns a vector as in FALSE, but with pre-weighted values, such that sum(divergence(..., summed = 'weighted)) is equivalent to divergence(..., summed = T).

na.rm
logical. Should missing values (including NaN) be removed? Used only if summed is set to TRUE.

Source

Wendell Bell, A Probability Model for the Measurement of Ecological Segregation, Social Forces, Volume 32, Issue 4, May 1954, Pages 357–364, https://doi.org/10.2307/2574118

Value

A scalar value, see note.

Details

This will report the percentage of population white in the geographic unit, e.g., tract, for the typical or average white person. The maximum value of this isolation index is 100. Even if whites make up only 20 percent of a metropolis population, all of them could live in all-white neighborhoods. The minimum value of the isolation index is asymptotically close to 0. That is, if there is only one white person in a metropolis of 100,000, he or she would live in a geographic unit in which the percent white was close to zero.

Note

Setting summed == FALSE will return by-observation measures, but this measure is not meant to be decomposed. These results are for verification purposes only.