Location Quotient

A location quotient (LQ) is an analytical statistic that measures a region’s concentration of a particular attribut relative to a larger geographic unit (e.g. census tract to state). In economics, an LQ is computed as an industry’s share of a regional total for some economic statistic (earnings, GDP by metropolitan area, employment, etc.) divided by the industry’s share of the national total for the same statistic. For example, an LQ of 1.0 in mining means that the region and the nation are equally specialized in mining; while an LQ of 1.8 means that the region has a higher concentration in mining than the nation.

location_quotient(group, population, na.rm = FALSE)

lq(group, population, na.rm = FALSE)

Arguments

group
A numeric vector of the proportion of the population

population
A numeric vector the length of group with its proportion of population.

na.rm
logical. Should missing values (including NaN) be removed? Used only if summed is set to TRUE.

Source

U.S. Bureau of Economic Analysis: https://www.bea.gov/help/faq/478

Details

LQ is equivalent to the percentage of a group in a single observation divided by the percentage of that group in the population of all observations, i.e. \( (group/population)/(\Sigma(group)/\Sigma(population))\)

Examples

data("bay_race")
# calculate the location quotient for Hispanic people in each tract in the bay area
location_quotient(group = bay_race$hispanic,
population = bay_race$total_pop)

#>    [1] 0.16937858 0.32524182 0.34871179 0.38120707 0.41103909 0.34101300
#>    [7] 0.44378419 0.43417476 0.56703470 0.60355414 0.53403724 0.39772976
#>   [13] 0.58812254 0.89787280 0.65636719 0.88260503 1.08886557 0.82140592
#>   [19] 1.28357335 0.39787498 0.19304864 0.24380645 0.65850626 0.35082420
#>   [25] 0.40620755 0.05490177 0.60984789 0.25380969 0.39687912 0.42771106
#>   [31] 0.51679783 0.41361018 0.38787533 0.36747313 0.35626485 0.38318729
#>   [37] 0.45851662 0.39774778 0.27749613 0.22827834 0.22461536 0.28404254
#>   [43] 0.24593212 0.25508080 0.23344566 0.25458797 0.48791569 0.41087015
#>   [49] 0.31048170 0.22895285 0.40141334 0.60273828 0.62181011 0.89076893
#>   [55] 0.83971727 0.52403917 0.55570706 0.73289446 0.78173745 1.70073469
#>   [61] 0.98625801 0.90459367 2.45443850 2.08979284 2.78649410 1.62713129
#>   [67] 0.73738147 1.86994775 0.99357542 1.26037198 0.37228840 0.47132379
#>   [73] 0.53962542 1.11696509 2.38620948 2.04823024 3.17910828 2.68235581
#>   [79] 2.75928504 2.10374521 1.47935982 0.57739715 0.74341393 0.55622698
#>   [85] 0.40910597 0.32930266 0.85686991 0.68976281 1.60963304 1.93521852
#>   [91] 1.91688779 1.90910832 1.95307012 2.52694104 2.07444624 2.20416724
#>   [97] 2.10298650 2.55565554 2.85463538 2.88313208 2.50004304 1.65691252
#>  [103] 0.54197295 0.34446561 0.40623609 0.75458615 1.44134765 2.62853933
#>  [109] 1.84335959 0.38776441 0.37011742 0.48416360 0.39253448 0.57707387
#>  [115] 0.47713419 0.32507189 0.21344562 0.22792613 0.22343159 0.16548266
#>  [121] 0.21960036 0.22466141 0.30978047 0.21185036 0.38748842 0.53753289
#>  [127] 1.20508296 0.48414680 0.39041075 0.34756435 0.29483539 0.49342303
#>  [133] 0.43432273 0.54723090 0.47558598 0.43234100 0.68754474 1.34981359
#>  [139] 0.63188799 0.66976638 0.34772879 0.32669298 0.37867569 0.32184286
#>  [145] 0.18606328 0.46517620 0.36710446 0.85814789 0.72234243 0.14918725
#>  [151] 0.31354834 0.41634692 0.56110853 0.13856881 0.20430322 0.39649602
#>  [157] 0.61287737 0.44637338 0.57807691 0.50592139 0.50307292 0.52135506
#>  [163] 0.75946400 0.56587550 0.41913776 0.34661004 0.23140474 0.56404903
#>  [169] 0.39673123 0.34966190 0.51199588 0.33894758 0.43541785 0.53240176
#>  [175] 0.53756192 0.49459293 0.98687035 0.59440490 0.71569984 0.79692379
#>  [181] 1.14357241 0.96504474 1.17996709 0.99438726 0.96765939 1.04941166
#>  [187] 1.51133801 1.58913677 1.54819928 1.36798225 1.28512107 0.90252870
#>  [193] 0.87339955 1.20498295 1.21618092 1.69712123 1.40672470 0.97403248
#>  [199] 1.06202594 0.48958501 1.02681075 0.99147882 1.93884824 1.50709696
#>  [205] 2.01541829 1.94658209 0.64939232 0.51861385 1.52596594 0.93355545
#>  [211] 1.36601765 1.40594823 1.77835211 2.44510969 2.23163583 1.91489991
#>  [217] 1.61083149 1.20606999 1.70465343 1.70606868 2.29756946 2.39519937
#>  [223] 1.21601806 0.70809809 1.54839963 2.44388120 1.88453756 2.47160667
#>  [229] 1.91211232 2.65771412 1.34545596 0.95481693 1.35026572 1.15375525
#>  [235] 1.50474839 2.37294409 2.54002172 1.85299194 2.35721175 2.96809373
#>  [241] 1.60007446 2.55874006 1.39224017 1.88133388 2.18866573 1.06295918
#>  [247] 1.23404650 1.88489358 1.38991831 0.69759572 3.15781762 1.46651152
#>  [253] 0.54489735 0.68874718 0.46824445 0.88382574 1.37246948 0.78403604
#>  [259] 0.26590036 0.33547426 0.69558185 0.36064686 1.01552054 0.71850908
#>  [265] 0.70801565 0.40793023 0.65695586 0.64351129 0.47450270 0.28448347
#>  [271] 0.17673250 0.25428432 0.59081747 0.26384957 0.16858904 1.02264084
#>  [277] 1.05014491 0.97661876 0.67167241 1.14798431 0.72292603 0.58411070
#>  [283] 0.35276645 1.36725730 0.50155715 0.15222952 0.11604897 0.25851167
#>  [289] 0.84106524 0.90222768 1.22258155 1.00772741 0.79860201 1.02734565
#>  [295] 0.71628575 1.01162308 0.81325095 1.49404616 0.98141415 0.17981092
#>  [301] 0.10236490 0.22414325 0.11831862 0.12971316 0.28700420 0.32477429
#>  [307] 0.34470502 1.22649136 1.51781216 1.43823478 2.05463125 2.25229286
#>  [313] 1.69667081 1.11631924 0.56669971 0.43114739 0.91220241 0.63981413
#>  [319] 0.63716824 0.89553599 0.50685046 0.42365972 0.29444504 0.38267505
#>  [325] 0.47384652 0.45124789 0.30691684 0.26624967 0.83158185 0.28222718
#>  [331] 0.76690008 0.44228022 0.64936163 0.38285013 0.30268280 0.38444590
#>  [337] 0.49005415 0.28747643 0.30624214 0.65870144 0.38585604 0.76655555
#>  [343] 0.87517168 0.89314497 0.94412580 1.11275547 2.19152915 0.71501539
#>  [349] 0.92248330 0.72413654 0.44340401 1.43760317 0.58569811 0.66264165
#>  [355] 0.83062961 0.59937786 0.74161054 0.44756176 0.65873527 0.45667876
#>  [361]        NaN 0.69372684 1.66345372 1.35352198 1.70756764 1.65159452
#>  [367] 1.09744270 1.20367765 1.45445831 1.76019434 1.01305150 1.02392396
#>  [373] 0.89934033 1.17937126 0.78957565 0.72731795 1.65481724 0.51522402
#>  [379] 0.47773446 0.92003235 1.93837511 1.40445930 1.56885022 1.29064478
#>  [385] 1.19716862 2.10228767 1.95729201 1.96429884 1.70172349 1.32100280
#>  [391] 1.28160910 1.09083842 1.24133378 2.28231902 2.22751694 1.62875770
#>  [397] 1.66531738 1.50906052 1.39562425 1.68648272 1.91340325 1.99027392
#>  [403] 2.42392359 2.31004737 2.31681402 2.18565195 2.93929040 0.97840193
#>  [409] 0.88878583 0.64650129 0.65072493 0.77671167 1.41492658 0.49566867
#>  [415] 0.63750190 0.64404062 0.44622356 0.43854713 0.61168661 0.48213103
#>  [421] 0.52170974 0.57305362 0.71456661 0.46343015 0.33277395 1.35715834
#>  [427] 0.93201156 1.45434517 1.20017285 1.00581736 0.92553919 0.85015948
#>  [433] 0.75110267 0.83659157 0.86533780 0.90207965 0.50840087 0.26347119
#>  [439] 1.13829697 2.96353551 2.76327400 2.32522321 3.28079806 0.74407143
#>  [445] 0.91334875 0.39394683 1.79060538 0.67538201 0.44088967 0.40524737
#>  [451] 0.31464493 0.23572811 0.27164309 0.72086464 0.53777267 0.64173189
#>  [457] 0.31645813 0.38112681 0.61928542 0.29316317 0.24561742 0.58397827
#>  [463] 0.54362380 0.41073650 0.37441853 0.33422571 0.39858363 0.33946800
#>  [469] 0.24901234 0.23055782 0.35147726 0.47292203 0.41234548 0.34358940
#>  [475] 0.21223613 0.22293347 0.24833737 0.21137804 0.25923132 0.25412874
#>  [481] 0.28042417 0.18631332 0.33571908 0.30469440 0.08359432 0.11579968
#>  [487] 0.05067194 0.19386386 0.43452391 0.31131422 0.27285700 0.16520088
#>  [493] 0.22339954 0.20852559 0.15781427 0.15936323 1.13371082 1.01538375
#>  [499] 1.00397953 1.07992803 1.18110300 0.26368572 0.22517250 0.30015232
#>  [505] 0.46317555 0.18466664 0.18521547 1.00070096 0.58080752 0.26848567
#>  [511] 0.40992679 0.42069169 1.01799762 0.60198098 0.63678933 1.04570465
#>  [517] 1.05327531 0.79127431 0.55898889 0.52141635 0.91496096 0.77438299
#>  [523] 0.65920650 0.80093603 0.65163378 0.90900703 0.87412062 0.63155218
#>  [529] 1.01179587 1.59424614 2.29406130 0.88162577 2.38308192 2.59450664
#>  [535] 1.25662246 1.94050177 3.33314626 3.00026786 1.73245669 1.19311745
#>  [541] 0.81679627 1.79799265 2.46390266 3.26660765 2.13724150 3.34121323
#>  [547] 2.38122825 2.68472770 0.61023450 2.24373532 0.97316838 1.81175526
#>  [553] 1.08005445 0.74987800 0.35149101 0.33522619 0.24158163 0.74218072
#>  [559] 0.47690428 0.51167290 0.61802422 0.85830543 0.23843858 0.33144044
#>  [565] 0.24619399 0.19429038 1.57764993 0.35911709        NaN 0.29128971
#>  [571] 0.77205238 0.37794104 0.80774060 1.41340249 0.35412200 0.89168286
#>  [577] 0.80860484 1.45647143 0.95261354 0.25288702 0.99618147 0.63231366
#>  [583] 0.86472910 0.30736169 0.28872135 0.82349793 1.32200893 0.28266761
#>  [589] 0.70149127 0.17479862 1.00992021 1.11050289 3.84565537 2.61240430
#>  [595] 0.27446211 0.25264177 0.32091933 0.20939519 0.22851891 0.25734554
#>  [601] 0.16427952 0.21494423 0.47328839 0.14801098 0.23774166 0.39355165
#>  [607] 0.38942907 0.83915525 0.14760475 0.22583731 0.18893693 0.26436807
#>  [613] 0.16983921 0.24316223 0.18663883 0.22354007 0.19579606 0.62362675
#>  [619] 0.17721538 0.16327511 0.27854036 0.53906887 0.62794883 0.75257246
#>  [625]        NaN 1.78120800 1.54233293 2.44008091 2.07209236 1.50614184
#>  [631] 0.77327489 2.30276016 2.00934737 2.03567179 1.85974647 1.53679824
#>  [637] 1.26184769 0.42978855 1.76798793 1.68940054 0.93895267 0.95504454
#>  [643] 1.42736354 0.80126992 2.97307735 0.78123179 0.77499872 1.09145159
#>  [649] 1.08845613 1.11767746 1.35904881 0.35787002 0.55242842 0.99589339
#>  [655] 0.41895016 0.37957951 0.28730628 0.75470102 1.21090853 2.04318695
#>  [661] 0.98249133 0.75572382 0.76514227 0.85497103 2.09911241 0.32522741
#>  [667] 0.21654298 0.21855235 0.22358393 0.24545054 0.27734117 0.12697724
#>  [673] 0.15417320 0.26180291 0.19907087 0.41168055 0.24325785 0.10010688
#>  [679] 0.40777570 0.10771320 0.38845684 0.51879609 0.64115879 0.55130781
#>  [685] 1.18794475 0.78315814 0.93310243 0.51470331 1.55160623 0.83132288
#>  [691] 0.81450368 0.89688050 0.23002637 0.19585635 0.23012518 0.20399704
#>  [697] 0.26438875 0.29596612 0.23866640 0.28338140 0.21704614 0.16243201
#>  [703] 0.19490934 0.21501369 0.27114608 0.25582652 0.28316757 0.30012965
#>  [709] 0.27609329 0.40734051 0.30080810 0.44570454 0.47203196 0.43163671
#>  [715] 0.34601154 0.41052217 0.41361513 0.56557293 0.54472634 0.54020160
#>  [721] 0.38645457 0.37326245 0.42408690 0.51529407 0.55232193 0.42023968
#>  [727] 0.32848291 0.32272688 0.29939042 0.58009987 1.44658355 0.16890692
#>  [733] 0.60821225 0.97880825 0.60639794 1.82969513 1.06957194 0.65164933
#>  [739] 0.31522753 0.47762689 0.31604826 0.41516555 0.77208948 2.16309042
#>  [745] 2.12689088 0.92616272 0.38622965 0.33353739 0.33684721 0.51193990
#>  [751] 0.81599221 0.41454300 0.54817246 0.61230878 0.36586861 0.52425560
#>  [757] 0.33911190 1.79434543 1.34097947 2.24389642 2.55568375 2.27782770
#>  [763] 2.14979097 0.63989491 0.99933226 0.69070953 0.68942534 1.65173928
#>  [769] 0.89444126 1.42308743 0.99352311 1.17204663 1.21648148 1.59631238
#>  [775] 1.20166951 1.58833554 1.35269743 1.18972719 0.95284478 1.04359122
#>  [781] 0.92629300 0.79528605 2.04070758 1.42407737 1.65464714 1.59727169
#>  [787] 1.12882831 1.18999714 1.19400297 1.07317994 1.14634227 0.83690950
#>  [793] 0.90343787 0.68912619 1.23781988 0.33247294 0.26804675 0.31174363
#>  [799] 0.33715005 0.29608054 0.27075658 0.23066173 0.43690903 0.33693078
#>  [805] 0.40621363 0.27535276 0.30947117 0.45146864 0.76344945 0.67359242
#>  [811] 0.73326225 0.31783202 0.54473031 1.09416649 0.28210783 0.17535379
#>  [817] 0.21867328 0.24880970 0.16389586 0.21170268 0.20643217 0.27655558
#>  [823] 0.23252232 0.75962763 0.60495603 0.67655502 0.23205638 0.23184544
#>  [829] 0.37383109 0.24007049 0.31455077 0.33698510 0.32025971 0.32174192
#>  [835] 0.23992788 0.27462107 0.14136999 0.28595219 0.26138879 0.25989424
#>  [841] 0.28940744 0.20569817 0.23674397 0.25120261 0.24913106 0.28497888
#>  [847] 0.32858089 0.44557436 1.10791019 0.35295441 0.70078749 0.08983449
#>  [853] 1.56181432 0.71875349 0.27465786 0.33217474 0.47242630        NaN
#>  [859] 0.81088981 0.54627453 0.98399650        NaN 0.76557587 1.29274670
#>  [865] 0.87739665 1.04458705 1.64183987 1.30245704 1.82164828 1.52870609
#>  [871] 0.88202996 0.37188707 0.72022288 0.52813227 1.30846671 1.71828555
#>  [877] 0.47512264 0.43584895 1.15152183 1.43736478 0.48456277 0.70976876
#>  [883] 0.40945140 1.31047044 1.14179935 1.33676343 1.40848218 1.48750360
#>  [889] 3.07920015 2.83227867 1.92822438 1.05547916 0.40080158 0.45813840
#>  [895] 0.83495948 0.77665129 0.76944315 0.73994069 0.56344951 0.61422824
#>  [901] 0.65729518 0.69248408 0.84566918 0.81167980 0.72237495 0.98329759
#>  [907] 0.86516804 2.19214552 0.73825515 2.41957039 0.88898565 0.45117962
#>  [913] 0.30592284 0.38680859 0.55553747 0.21370626 0.35343428 0.63199166
#>  [919] 0.36702286 0.53309643 0.90473441 0.64461098 0.16606346 0.12578921
#>  [925] 0.29795068 1.34589418 1.88113123 1.75540974 2.65012100 1.18379152
#>  [931] 0.44106190 0.37893480 0.90019984 0.34405834 0.29712897 0.39764141
#>  [937] 0.44997840 0.35739649 1.07688883 0.39397864 0.83871324 0.95303821
#>  [943] 1.09540138 1.53018808 1.57938111 1.11543438 0.45431316 0.29824260
#>  [949] 0.21156424 0.25298461 0.25622330 0.24353400 0.31137493 0.37177100
#>  [955] 0.28753968 1.39453498 1.12789456 0.80618823 0.74619296 0.48773579
#>  [961] 0.29559820 0.48316326 0.47131992 1.02044081 0.50203556 0.56440310
#>  [967] 0.37708597 0.29832900 0.46577379 0.30184575 0.34177932 0.30941347
#>  [973] 0.32475269 0.73356488 0.57047402 1.02002792 2.03177544 3.16527262
#>  [979] 1.72244024 3.57867049 1.71243329 0.26834793 0.43332166 2.96411168
#>  [985] 3.43882180 3.50299158 1.87477620 1.84379630 2.86127627 2.39641172
#>  [991] 1.42522580 0.53354690 0.91545370 0.75437254 0.18865832 0.12706273
#>  [997] 0.45864198 2.91646087 2.68173595 2.41767878 3.09414404 2.56644029
#> [1003] 0.31225866 0.28338758 0.17957349 0.15608988 0.20497835 0.21598226
#> [1009] 0.19576520 0.19362018 0.21582230 1.72496636 0.63119317 0.86972599
#> [1015] 1.42488737 1.09454274 0.39966143 0.69098189        NaN        NaN
#> [1021] 2.46028350 1.67458976 1.64386299 1.58838138 1.07605726 1.15799042
#> [1027] 1.76941821 0.83711961 0.99269060 2.23245946 1.63957670 2.74655166
#> [1033] 2.66120693 1.12617876 3.21049158 2.92737588 2.83753947 2.97787465
#> [1039] 2.09862323 3.57116229 2.30276467 1.62157962 2.15639748 2.96854244
#> [1045] 1.00934598 2.16149739 1.46912505 0.58370163 0.91588101 1.21115660
#> [1051] 0.81379643 0.74907096 0.48897741 0.62927857 0.97536347 0.84879711
#> [1057] 0.58844743 0.82922218 0.85122935 0.74997676 0.68615848 0.98090372
#> [1063] 0.67603334 0.86779968 1.22450886 1.14087851 0.55567937 0.95684517
#> [1069] 0.90223325 1.34025288 1.40902012 2.72253815 2.03631388 2.54159482
#> [1075] 3.38807073 0.83809226 1.02319964 2.49386577 2.32942242 2.39951399
#> [1081] 1.32030070 1.97351640 2.36698580 0.83892258 2.76449409 1.97935269
#> [1087] 1.62760023 1.33904567 1.42768616 1.50540696 2.83644249 3.28278693
#> [1093] 1.86586531 2.05691096 2.64420651 1.59965090 1.01888187 1.48282877
#> [1099] 0.54859465 1.63939738 1.36769833 0.73924116 1.31627229 0.47889108
#> [1105] 1.41126329 0.60810725 0.31980867 0.87636345 0.36865636 1.14618761
#> [1111] 0.31214071 1.57069570 0.98180005 3.45596901 2.93985062 3.40054836
#> [1117] 2.88413614 3.31845918 3.19485979 2.68160395 3.28363195 2.00816627
#> [1123] 2.31207832 2.53262748 3.12093377 2.05819412 1.02044081 1.50084488
#> [1129] 3.01137311 3.42610736 1.91270552 2.63059165 1.83038227 0.81104949
#> [1135] 1.16455930 2.71758020 2.12012216 3.03500727 2.82032472 2.67966213
#> [1141] 3.04456338 1.44930781 1.37598071 0.65042478 0.70205866 0.32933371
#> [1147] 0.64855494 0.71598287 0.72401601 0.61727952 1.07810105 0.78158763
#> [1153] 0.49504678 0.85100361 0.44756176 0.58285935 0.91446635 0.77313215
#> [1159] 1.12520552 0.16868430 0.28807613 0.49034083 0.51741928 0.61868832
#> [1165] 1.34965122 0.47512292 0.53271481 0.88000978 0.89050268 0.88203454
#> [1171] 0.71698016 0.43063843 0.85655934 1.08767916 2.59790397 0.97573444
#> [1177] 1.06255363 0.73646303 0.70837174 1.56868106 0.31899929 0.65386012
#> [1183] 0.22360865 0.97262197 0.35607312 1.01674356 1.37513979 1.76755237
#> [1189] 0.89651953 1.57521709 1.05157506 1.23618216 0.88813251 0.77032733
#> [1195] 0.68591826 1.05523112 0.80394851 0.87406718 0.82735483 1.03283184
#> [1201] 0.96834132 0.82646240 0.81132080 0.61819771 0.49909953 0.42436932
#> [1207] 0.35681458 0.54777841 0.55111208 0.99719680 0.83199685 1.28892912
#> [1213] 1.72474033 1.14935495 1.29416624 2.20919039 0.66928067 1.04449587
#> [1219] 0.90026619 0.50666850 0.50680686 1.01629794 0.89562606 0.51293023
#> [1225] 0.77398238 0.60757281 0.38383128 0.56425941 0.81478609 0.29290691
#> [1231] 0.29504408 0.22030581 0.17766603 0.36290369 0.34408861 0.28218894
#> [1237] 0.20291438 0.15437638 0.18567654 0.10281824 0.30894562 0.08123130
#> [1243] 0.10853151 0.11895458 0.07565299 0.14235321 0.39233179 0.11912940
#> [1249] 0.08981823 0.10652750 0.08894687 0.08965857 0.18598110 0.10867923
#> [1255] 0.14697280 0.30807638 0.31347696 0.13058312 0.18180091 0.51183615
#> [1261] 0.20552188 0.37611622 0.18993466 0.15765983 0.52313034 0.34654274
#> [1267] 0.16150037 0.34873109 0.51824570 0.79670178 0.38045807 0.62404649
#> [1273] 1.11448933 0.83344346 0.77195368 0.65697953 0.97368444 1.70028309
#> [1279] 1.77475680 2.01312156 0.44840129 0.59429988 1.10908616 1.44625027
#> [1285] 1.12826127 0.53319167 0.78145859 0.93293650 0.55396664 1.03599799
#> [1291] 1.22179216 0.67775230 2.12301240 1.34356921 1.65030859 0.53775611
#> [1297] 0.74760556 0.55707607 0.43159505 0.18906823 0.34450599 0.13834459
#> [1303] 0.19865041 0.20913037 0.16504357 0.15441994 0.20369706 0.10883541
#> [1309] 0.43771963 0.45056113 0.15833872 0.18717325 0.24932584 0.29700914
#> [1315] 0.32675784 0.12493919 0.16221691 0.28492286 0.22504798 0.19979777
#> [1321] 0.27334449 0.51110654 0.20704829 0.08996916 0.11881064 0.12793137
#> [1327] 0.19656733 0.31793395 0.35847364 0.30030550 0.35396434 0.19512618
#> [1333] 0.22922460 0.48127324 0.24151058 0.41676909 0.32614370 2.14490975
#> [1339] 0.77132270 0.99360164 1.06058332 1.97990449 1.36073136 1.54251528
#> [1345] 1.12390427 1.61639420 1.99878703 0.98511506 1.01417523 1.44876586
#> [1351] 1.26462218 1.36525169 0.99725904 1.11265723 1.45784428 1.15997048
#> [1357] 1.21618161 1.16079627 0.85871337 1.29380719 1.14099166 1.36513821
#> [1363] 1.76897959 2.69581294 1.09597505 1.19979498 0.96167420 1.44885041
#> [1369] 1.78903814 0.85587545 1.01326800 0.99995908 1.00439477 1.06891652
#> [1375] 2.07040111 1.57294865 1.36839036 2.35180061 2.05352532 1.83428255
#> [1381] 1.73134594 1.37086788 2.61055198 3.07599973 2.47680847 2.29915666
#> [1387] 1.78338260 2.35787542 3.66040774 3.54303997 0.46840076 1.27852433
#> [1393] 0.73703787 0.55758736 0.43868157 0.38258821 0.94922255 1.01322622
#> [1399] 1.14594368 1.12301047 0.59128804 1.49153889 0.51017614 0.51750984
#> [1405] 1.52733095 0.55576727 0.81298522 1.14988945 1.77950611 1.60294661
#> [1411] 0.80463622 0.93259439 1.70432274 1.20399616 1.04026254 0.68939789
#> [1417] 1.05877100 1.61617531 0.88293955 0.90032978 1.47607910 1.10381142
#> [1423] 0.43878710 0.68485961 0.63120897 0.42088268 0.45182294 0.46819596
#> [1429] 0.54450025 0.60049637 0.52365390 0.86681211 0.75903092 0.78780626
#> [1435] 0.46613926 0.62513290 0.81047481 0.64302210 0.95105514 0.82946143
#> [1441] 1.10048998 0.68547859 2.02566119 1.79554464 1.78831934 1.42600062
#> [1447] 1.56348160 1.95646110 1.88640153 2.10977496 1.70765914 1.10152613
#> [1453] 1.31086354 0.86435694 1.14642607 1.17863718 1.05618420 0.95787997
#> [1459] 0.90260956 0.64946589 0.70081783 0.74757569 0.85198675 1.02015939
#> [1465] 0.69133419 0.88639858 0.50983954 0.57371450 0.83737803 0.93821388
#> [1471] 0.63910951 0.85690026 1.12151383 1.28739736 1.11197914 1.29972203
#> [1477] 1.06503934 0.52505177 1.12883343 1.53528565 1.52373897 0.71275148
#> [1483] 1.86008169 2.03034393 1.34954446 1.67668104 0.73295345 2.39165816
#> [1489] 0.70497723 0.73488536 0.58987746 0.61153388 0.39357529 1.70681071
#> [1495] 2.56462548 1.54979293 0.49737556 1.39856209 0.83220301 1.40007370
#> [1501] 0.60027794 1.03796238 0.73058533 1.32980976 0.65571678 0.85331937
#> [1507] 0.31758007 0.62747370 1.31699429 0.42010685 0.62746026 0.61660300
#> [1513] 0.83548126 0.78489072 0.79106836 1.36166299 1.53239185 0.82089917
#> [1519] 0.65084165 0.61614398 0.64236382 0.70460195 0.69553706 2.16840181
#> [1525] 2.10252098 0.56123471 0.20830667 0.46062559 0.24622839 0.13047152
#> [1531] 0.76591556 0.50836293 1.06093963 1.10437317 0.96671254 0.77828733
#> [1537] 0.76134471 0.50149251 0.26911602 0.28356915 0.71542394 0.66674653
#> [1543] 0.30446506 0.87198549 1.00090929 0.89375187 1.83952683 2.24148451
#> [1549] 1.02895283 1.15011288 0.83103343 1.81705731 2.26013117 1.79130453
#> [1555] 1.03984791 1.23481835 2.82419109 2.51630116 2.77220595 2.09256841
#> [1561] 1.61152722 0.52567154 0.37867921 0.40493683 0.46367564 0.43741751
#> [1567] 0.60421258 0.32528378 0.54288473 0.47977456 0.46342111 1.52480877
#> [1573] 1.14467478 1.13796117 0.67945694 1.84336276 1.22377588 1.09884214
#> [1579] 2.11720348 0.83001281 0.76948913 1.25878995 1.84372640 1.00798465
#> [1585] 0.35290131 0.39304770 0.52262160        NaN