Divergence Index

Elizabeth Roberto's Divergence index for calculating and decomposing segregation.

divergence(
  ...,
  population = NA,
  na.rm = TRUE,
  summed = FALSE,
  logBase = exp(1),
  comparison = NULL,
  sumPercent = NA,
  weights = NA,
  rowTotals = NA
)

Arguments

...
Vectors containing proportions of each group, or a dataframe containing only those vectors.

population
A vector of population totals for each row, or weights summing to a total of one. If NA, will assume all populations/weights are equal. Set to 1 to silence warning. If a string, will use the string as the named column of the dataframe provided in ...

na.rm
logical. Should missing values (including NaN) be removed? Used only if summed is set to TRUE.

summed
If TRUE, will return a single summary statistic. (Or one value per group if specifying dplyr::group_by.) If FALSE (default), will return a vector equaling the length of the input vectors. If 'weighted' (only for divergence and information theory), returns a vector as in FALSE, but with pre-weighted values, such that sum(divergence(..., summed = 'weighted)) is equivalent to divergence(..., summed = T).

logBase
Specify the base for the logathirm used in the equation. Natural logarithm by default.

comparison
A vector of percentages that must equal the length of the number of vectors or columns in ..., representing the percentages of each group in the larger (comparison) geography.

weights
deprecated, use population.

rowTotals, sumPercent
deprecated, will throw error.

Source

Created by Elizabeth Roberto: https://arxiv.org/abs/1508.01167

Value

A single value if summed==TRUE, or a vector equaling the length of the inputs.

Details

The demographics of each observation are compared to that of their larger geography, which is inferred from the combination of the percentages in ... and the population totals in population, or can be directly provided with comparison.

Examples

library(rsegregation)
data("bay_race")
#return by-observation scores
divergence(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, bay_race$all_other, population=bay_race$total_pop)

#>    [1] 0.219578344 0.293020051 0.198434902 0.208270938 0.332195755 0.551009770
#>    [7] 0.731421349 0.504571655 0.605739288 0.746098754 0.213119375 0.199658530
#>   [13] 0.379173063 0.723961096 0.817288048 0.658662858 0.362408493 1.007226116
#>   [19] 0.482917645 0.978072828 1.292655290 0.601956246 0.924319984 0.534803074
#>   [25] 0.266179287 1.018730202 0.363482918 0.299307929 0.174139421 0.265582946
#>   [31] 0.252894004 0.449823243 0.191288226 0.206557541 0.139370531 0.152429140
#>   [37] 0.139578404 0.154601028 0.175831649 0.229199178 0.236267624 0.165056060
#>   [43] 0.223153949 0.232858179 0.236774225 0.229312183 0.140079512 0.084212125
#>   [49] 0.170669595 0.240664649 0.115653811 0.173224464 0.193518531 0.317820113
#>   [55] 0.420655244 0.331915305 0.261175669 0.471071823 0.477364783 0.442880447
#>   [61] 0.435072224 0.220116125 0.324957164 0.482983805 0.548926588 0.378428440
#>   [67] 0.243698584 0.361610494 0.348144286 0.345728566 0.099688096 0.091608649
#>   [73] 0.152834934 0.327208648 0.445189030 0.409177843 0.646816528 0.454501837
#>   [79] 0.697828980 0.849094832 0.538414484 0.762325765 0.341236304 0.165972081
#>   [85] 0.093973262 0.313557551 0.864374100 0.618147044 1.106872822 1.082622656
#>   [91] 1.021274232 0.945275370 0.920024165 0.940212096 1.047143805 1.051645531
#>   [97] 0.970021021 0.914214016 0.847351422 0.923939095 0.958890099 1.055637270
#>  [103] 1.204082079 0.827587406 0.722546842 1.124612565 1.073986057 0.890100596
#>  [109] 0.816635672 1.141915679 0.120057124 0.074003346 0.101241496 0.168928767
#>  [115] 0.058944361 0.177483374 0.274958887 0.315421493 0.300000772 0.363745801
#>  [121] 0.320567270 0.289034000 0.167682900 0.270732667 0.126870506 0.278404772
#>  [127] 0.089813115 0.094603412 0.098295933 0.094490605 0.130657402 0.116450493
#>  [133] 0.085269081 0.130211047 0.084630405 0.096326682 0.107288184 0.204029259
#>  [139] 0.373081229 0.218443523 0.171153912 0.173252182 0.130541014 0.152522775
#>  [145] 0.389234856 0.232807640 0.212134921 0.322322881 0.541363881 0.165514612
#>  [151] 0.161381007 0.119343580 0.193321286 0.265284676 0.222535443 0.161537445
#>  [157] 0.050349768 0.068946689 0.183730838 0.047735812 0.055682080 0.057733457
#>  [163] 0.037687455 0.068720818 0.078188701 0.149902036 0.188090838 0.033282053
#>  [169] 0.077051248 0.107874559 0.100891246 0.169969906 0.140315723 0.104209504
#>  [175] 0.114446243 0.122994732 0.128910656 0.038735792 0.057341669 0.041507097
#>  [181] 0.028153125 0.006933377 0.076046675 0.055563554 0.065586561 0.141770126
#>  [187] 0.105419018 0.085813021 0.128896590 0.200825978 0.133661218 0.023214369
#>  [193] 0.019190691 0.012589244 0.106429797 0.325764773 0.206180630 0.112490950
#>  [199] 0.086504602 0.237722543 0.051033953 0.086351258 0.175048468 0.178573858
#>  [205] 0.489109205 0.260727237 0.087517528 0.078456651 0.215917031 0.196490243
#>  [211] 0.205543361 0.157491505 0.162149257 0.364089480 0.241094758 0.121079393
#>  [217] 0.061005999 0.016427311 0.075774582 0.078281566 0.357481167 0.352134355
#>  [223] 0.087623523 0.070137591 0.187337676 0.370368717 0.309323096 0.332204198
#>  [229] 0.227829420 0.389996986 0.096639184 0.266951500 0.213827638 0.095343880
#>  [235] 0.193599387 0.272835679 0.439953106 0.177233333 0.487422007 0.622433077
#>  [241] 0.163851839 0.404303321 0.033307955 0.152876746 0.242192061 0.153610787
#>  [247] 0.083939284 0.317456191 0.077820339 0.042040054 0.627089321 0.148714303
#>  [253] 0.403303932 0.275027923 0.506383835 0.097794451 0.191005665 0.353419862
#>  [259] 0.687288745 0.578773697 0.319479896 0.523667996 0.167801948 0.027099561
#>  [265] 0.045588707 0.245400612 0.088199258 0.182172567 0.217716952 0.515109889
#>  [271] 0.759691508 0.464567453 0.299442249 0.630417240 0.730177287 0.027155219
#>  [277] 0.067951820 0.075141058 0.091345131 0.080082022 0.250026207 0.260227762
#>  [283] 0.328055929 0.139536740 0.282653042 0.537540034 0.619828245 0.401934259
#>  [289] 0.073906339 0.104208884 0.050552844 0.082582595 0.044135178 0.008918074
#>  [295] 0.051100899 0.036298689 0.067141221 0.076514161 0.091348962 0.551755099
#>  [301] 0.789754573 0.513619756 0.684348623 0.524840925 0.322771812 0.520936735
#>  [307] 0.312713231 0.020565056 0.061357323 0.047897257 0.189530648 0.217829006
#>  [313] 0.138690778 0.020764487 0.205037611 0.097195887 0.247996703 0.042579679
#>  [319] 0.064492857 0.049370592 0.117962692 0.095681116 0.157355812 0.116384596
#>  [325] 0.125077075 0.205229342 0.256425578 0.181534110 0.047058214 0.184676889
#>  [331] 0.106403095 0.135449312 0.128304694 0.116979632 0.174923433 0.213425794
#>  [337] 0.063246895 0.253168375 0.216284984 0.099567834 0.266232612 0.117250967
#>  [343] 0.146628606 0.177554823 0.107058915 0.130666345 0.237348631 0.142238121
#>  [349] 0.149166651 0.148739891 0.316580856 0.119295387 0.246190699 0.235810547
#>  [355] 0.191018816 0.161230398 0.198299296 0.143210351 0.102514211 0.096824845
#>  [361] 0.000000000 0.211267562 0.187077102 0.137714681 0.134401129 0.122670398
#>  [367] 0.201993250 0.117414804 0.059667167 0.168150295 0.058876741 0.127733793
#>  [373] 0.071107499 0.167802232 0.210661877 0.046110642 0.270018383 0.347034687
#>  [379] 0.349513584 0.144013284 0.274306737 0.155608174 0.191529212 0.168396628
#>  [385] 0.123012538 0.263452261 0.239857202 0.548142816 0.149609427 0.158252424
#>  [391] 0.131118357 0.084939105 0.392045062 0.387983343 0.388385418 0.667191322
#>  [397] 0.217994001 0.255047561 0.113977295 0.089907187 0.177803018 0.147885385
#>  [403] 0.408137401 0.283795674 0.343263537 0.296604490 0.484927372 0.007690398
#>  [409] 0.237669924 0.267612144 0.272329568 0.196301738 0.153410863 0.141084992
#>  [415] 0.122818933 0.146935514 0.189954460 0.264860731 0.044005854 0.170180999
#>  [421] 0.247035242 0.059970581 0.072223052 0.259381404 0.342014480 0.109337966
#>  [427] 0.011959035 0.110729060 0.103724898 0.075783123 0.133293491 0.095333444
#>  [433] 0.090754603 0.112760696 0.107098483 0.037295856 0.096673503 0.325059091
#>  [439] 0.129950060 0.512515097 0.406607718 0.236900030 0.658408749 0.097734393
#>  [445] 0.063963182 0.157696382 0.103540214 0.131894708 0.180605038 0.125186548
#>  [451] 0.185755674 0.265503876 0.255181366 0.043755142 0.131452733 0.122407262
#>  [457] 0.275301175 0.251867022 0.219304973 0.263144175 0.360984665 0.070931010
#>  [463] 0.133008027 0.143589221 0.271659068 0.127063441 0.108038426 0.108751580
#>  [469] 0.194080953 0.361911885 0.114725378 0.136994445 0.102073693 0.361265435
#>  [475] 0.452011599 0.304980405 0.329699479 0.413576394 0.371755725 0.271848249
#>  [481] 0.280408813 0.379120248 0.271607652 0.239601149 0.496888485 0.486318639
#>  [487] 0.550693434 0.387736084 0.145012559 0.239904621 0.207234788 0.305031568
#>  [493] 0.268808270 0.313698470 0.367365364 0.380024134 0.065452324 0.185197099
#>  [499] 0.075473173 0.159056327 0.142914394 0.259817962 0.177742151 0.148710878
#>  [505] 0.117059764 0.453278115 0.511009785 0.146275846 0.122927508 0.197218399
#>  [511] 0.256785166 0.343555931 0.076954405 0.073454510 0.238276863 0.094273252
#>  [517] 0.097389033 0.067861960 0.268644777 0.264371020 0.009939853 0.231241799
#>  [523] 0.299966280 0.016182125 0.052353501 0.068457368 0.052942434 0.079954576
#>  [529] 0.130806258 0.092323288 0.759035222 0.435148523 0.385730277 0.468068007
#>  [535] 0.675386727 0.315369467 0.750764945 0.566748882 0.312424859 0.110696502
#>  [541] 0.048183627 0.161745779 0.336796466 0.759130862 0.339364055 0.860021862
#>  [547] 0.809086396 0.700472164 0.157013969 1.036005994 0.237580378 0.916067319
#>  [553] 0.921714435 0.025350352 0.096483056 0.092840888 0.133277023 0.081866980
#>  [559] 0.059040077 0.049392528 0.032495827 0.058905498 0.158479319 0.162551991
#>  [565] 0.245886176 0.267581866 0.215824279 0.513229451 0.000000000 0.401041493
#>  [571] 0.225259931 0.331914045 0.171937932 0.133098209 0.373593702 0.162167790
#>  [577] 0.180205707 0.089537602 0.198721382 0.449148909 0.092015575 0.144275196
#>  [583] 0.154039066 0.330546864 0.355777920 0.134374373 0.175601346 0.409456809
#>  [589] 0.205310819 0.480204962 0.151327010 0.187775593 1.023934298 0.333734709
#>  [595] 0.496718809 0.452931518 0.404118951 0.470186036 0.488966630 0.391569831
#>  [601] 0.562479633 0.487129536 0.251604669 0.456135491 0.447943329 0.340923475
#>  [607] 0.250959504 0.609587485 0.547365186 0.382741323 0.418609910 0.244924718
#>  [613] 0.501189774 0.301154342 0.474455700 0.386389461 0.428112301 0.534575175
#>  [619] 0.458769615 0.464929295 0.408651991 0.392015366 0.383677581 0.368752836
#>  [625] 0.000000000 0.308423494 0.254974774 0.421651063 0.320809994 0.259092471
#>  [631] 0.330036108 0.387570870 0.348235652 0.306993944 0.200013827 0.267352794
#>  [637] 0.248696696 0.415534635 0.308464543 0.287043335 0.258875353 0.282586737
#>  [643] 0.240770990 0.346147177 0.595153665 0.230459035 0.150183051 0.053269838
#>  [649] 0.004848985 0.013469821 0.059386281 0.416955452 0.403650675 0.300820448
#>  [655] 0.413950267 0.429379630 0.486776644 0.330745773 0.335208845 0.338435156
#>  [661] 0.326841503 0.158513849 0.364382494 0.304502405 0.365688985 0.105947475
#>  [667] 0.338140526 0.193469217 0.191124352 0.176092547 0.290892540 0.724067247
#>  [673] 0.269181833 0.195202798 0.234733663 0.110681290 0.173445530 0.752137182
#>  [679] 0.153934636 0.970667436 0.085134085 0.071786059 0.030434954 0.041885809
#>  [685] 0.010418517 0.042747911 0.044837389 0.052984862 0.083725657 0.012791727
#>  [691] 0.071478799 0.161755001 0.331065993 0.358957119 0.343880760 0.356529943
#>  [697] 0.269378041 0.243575006 0.331035461 0.210730719 0.274964614 0.346108897
#>  [703] 0.331156328 0.309097832 0.220825537 0.145223170 0.125018589 0.159074360
#>  [709] 0.150933200 0.075209554 0.125619787 0.063123696 0.248884037 0.177864123
#>  [715] 0.137294287 0.074623871 0.439249176 0.074985242 0.137154583 0.128871203
#>  [721] 0.147272866 0.211482825 0.182493842 0.136473976 0.105260231 0.191978158
#>  [727] 0.290921667 0.286039599 0.248932156 0.097948215 0.043726454 0.428559913
#>  [733] 0.034727254 0.182427511 0.124936823 0.097050262 0.020904711 0.124648566
#>  [739] 0.240322288 0.126304106 0.271772510 0.233565391 0.120910091 0.184743168
#>  [745] 0.185354398 0.113082401 0.240119539 0.273195791 0.264612565 0.191246540
#>  [751] 0.135948552 0.089633954 0.057344145 0.164269979 0.162754840 0.161798741
#>  [757] 0.209906884 0.107626814 0.018884715 0.229882485 0.319571350 0.230131190
#>  [763] 0.182520613 0.554399017 0.448655839 1.135554392 1.294449907 0.718342736
#>  [769] 0.575004345 0.620560248 0.041637374 0.070592882 0.071794031 0.068828879
#>  [775] 0.007407124 0.071500688 0.189914365 0.315369978 0.378968856 0.433002865
#>  [781] 0.350100436 0.209462604 0.329427040 0.085202154 0.270576814 0.297735781
#>  [787] 0.396928170 0.418868172 0.461028572 0.287230230 0.373692401 0.536249013
#>  [793] 0.375566519 0.565878444 0.501402586 0.135640914 0.154282523 0.147868717
#>  [799] 0.120236871 0.161103824 0.215112686 0.193601121 0.068630529 0.131219080
#>  [805] 0.090058758 0.148036769 0.118732832 0.080796001 0.031639525 0.349069911
#>  [811] 0.379726294 0.312977010 0.350678433 0.318945199 0.253048551 0.354087615
#>  [817] 0.315366332 0.364529341 0.445614033 0.402968107 0.474262689 0.268042440
#>  [823] 0.299847143 0.014153371 0.030156576 0.036968906 0.354077925 0.282775147
#>  [829] 0.146976704 0.415943639 0.280369839 0.131686989 0.160584201 0.157435405
#>  [835] 0.167262065 0.193681920 0.266592679 0.197809587 0.198326254 0.224299810
#>  [841] 0.226320104 0.266937448 0.292904930 0.245182736 0.199097205 0.191421969
#>  [847] 0.269655119 0.069664583 0.547547081 0.118224776 0.361317139 1.030011755
#>  [853] 0.499606150 0.096708848 0.121905276 0.200995185 0.287223545 0.000000000
#>  [859] 0.481094012 0.971493064 0.089359793 0.000000000 0.040328356 0.414929049
#>  [865] 0.238193611 0.258118318 0.402706025 0.359429025 0.361777868 0.391037137
#>  [871] 0.215506921 0.203517958 0.259987285 0.315618394 0.104842306 0.313468128
#>  [877] 0.531266601 0.562677620 0.331330622 0.090934864 0.479610939 0.141848596
#>  [883] 0.604151517 0.101067216 0.056371552 0.105085549 0.123046105 0.124250336
#>  [889] 0.572728041 0.435976407 0.193291882 0.029824905 0.405455895 0.511983526
#>  [895] 0.100944896 0.095332551 0.042621887 0.142290680 0.173639252 0.221343832
#>  [901] 0.203387346 0.138941837 0.065855508 0.044354589 0.048453676 0.045657978
#>  [907] 0.065912342 0.249423635 0.086936324 0.414813961 0.090265512 0.124152994
#>  [913] 0.218798742 0.179493328 0.144089581 0.273837746 0.144951011 0.110958624
#>  [919] 0.220842467 0.141128947 0.074157581 0.090273955 0.236849747 0.242720870
#>  [925] 0.286037697 0.045347840 0.117507939 0.154390337 0.404512434 0.029966415
#>  [931] 0.114783871 0.240105515 0.061170612 0.194714590 0.156258347 0.112140387
#>  [937] 0.112246273 0.161839424 0.019131402 0.167007968 0.038378157 0.019440729
#>  [943] 0.055662394 0.114352227 0.133983382 0.054662177 0.088612522 0.199996082
#>  [949] 0.186949971 0.232773499 0.183743372 0.208970062 0.182815745 0.137326870
#>  [955] 0.204555416 0.046732186 0.063746380 0.034364476 0.031090911 0.130798953
#>  [961] 0.197677603 0.096774231 0.151527558 0.132420620 0.231345364 0.130686148
#>  [967] 0.296718307 0.337268564 0.183749387 0.222801102 0.237288421 0.279789068
#>  [973] 0.324846803 0.193254259 0.302296628 0.140367902 0.213049060 0.600977675
#>  [979] 0.128292163 0.838582595 0.125387171 0.174998641 0.084499709 0.511987363
#>  [985] 0.754247890 0.790119964 0.203493068 0.157818377 0.460493467 0.302090433
#>  [991] 0.130067883 0.232836170 0.190793965 0.183809531 0.290006589 0.343993040
#>  [997] 0.182300869 0.803507576 0.912958497 0.604740933 0.847153911 0.391551183
#> [1003] 0.234605321 0.269132684 0.418136625 0.428400190 0.294614324 0.251443536
#> [1009] 0.439938907 0.389104441 0.479260814 0.237280932 0.307014355 0.241085473
#> [1015] 0.223638275 0.265630307 0.183224858 0.110983375 0.000000000 0.000000000
#> [1021] 0.301598841 0.066362439 0.089010818 0.101849194 0.076111951 0.036493405
#> [1027] 0.081687428 0.027233507 0.028209563 0.203274522 0.075803345 0.381777163
#> [1033] 0.349174786 0.024943044 0.610207043 0.463901930 0.570569382 0.512055143
#> [1039] 0.161608943 0.824660533 0.291970081 0.058609777 0.194706183 0.491145882
#> [1045] 0.028797504 0.177596307 0.035305438 0.121651148 0.163641126 0.115005780
#> [1051] 0.187900158 0.201984387 0.234243304 0.183773625 0.045941259 0.071314354
#> [1057] 0.084672106 0.129356199 0.067755263 0.132851974 0.175144216 0.079736292
#> [1063] 0.124255875 0.052065602 0.079805274 0.053459564 0.226380995 0.099322909
#> [1069] 0.109586712 0.546039118 0.053730469 0.618586555 0.462535054 0.337843056
#> [1075] 0.706859955 0.024937570 0.044869139 0.597861184 0.400049688 0.266462758
#> [1081] 0.234374385 0.170307276 0.402070151 0.534889876 0.481783747 0.468077799
#> [1087] 0.345068902 0.412121220 0.476279807 0.470269551 0.582366602 0.710678398
#> [1093] 0.414135372 0.519812365 0.563809085 0.073565928 0.214360679 0.354834902
#> [1099] 0.742624673 0.213404583 0.282403945 0.535423486 0.368451655 0.674266496
#> [1105] 0.164732452 0.410218217 0.668809040 0.023950475 0.146958130 0.191354961
#> [1111] 0.327930761 0.398055499 0.418933709 0.824584166 0.626469141 0.816257728
#> [1117] 0.609305823 0.781467218 0.651888124 0.466002159 0.673283739 0.427237424
#> [1123] 0.231467417 0.441812096 0.729055116 0.440577614 0.552304659 0.615480285
#> [1129] 0.641059772 0.834202796 0.428879101 0.481447779 0.189957289 0.408928355
#> [1135] 0.256415594 0.404481807 0.318864915 0.584028192 0.584811363 0.383317232
#> [1141] 0.541827250 0.052816290 0.026713550 0.391372358 0.130729659 0.736803845
#> [1147] 0.400797875 0.314280552 0.495569402 0.496591500 0.247728653 0.467762342
#> [1153] 0.461193720 0.447629600 0.497967839 0.516170415 0.362531588 0.293259738
#> [1159] 0.353351428 0.679260182 0.695760844 0.403319807 0.369684172 0.103915010
#> [1165] 0.312109988 0.418003788 0.411685842 0.387569101 0.177357810 0.334278898
#> [1171] 0.374588229 0.490573296 0.270096099 0.048347508 0.335794524 0.125932862
#> [1177] 0.124953846 0.192898718 0.045650367 0.189854836 0.390762170 0.247994887
#> [1183] 0.404737823 0.118814389 0.316917147 0.053132057 0.070354835 0.105104675
#> [1189] 0.020445174 0.130773548 0.021102120 0.103187099 0.056749546 0.060827459
#> [1195] 0.196143034 0.043399500 0.064200474 0.034429797 0.032522253 0.037199437
#> [1201] 0.107906563 0.022074163 0.032051334 0.058994706 0.085309795 0.152997423
#> [1207] 0.130596113 0.284221670 0.144825338 0.002835417 0.023324812 0.036243526
#> [1213] 0.076414380 0.008747313 0.036713786 0.192331100 0.137682294 0.024416064
#> [1219] 0.010637883 0.112083248 0.066231056 0.012521714 0.011592099 0.082694002
#> [1225] 0.038620247 0.090665425 0.225065225 0.186257537 0.104307846 0.264138664
#> [1231] 0.276752919 0.393393907 0.408673162 0.295397444 0.199699090 0.285560187
#> [1237] 0.287727945 0.238009064 0.249666483 0.369890427 0.159325503 0.374622254
#> [1243] 0.279823632 0.440496129 0.473031532 0.308378394 0.283200962 0.670774663
#> [1249] 0.683818614 0.606742875 0.597050232 0.471556734 0.565907373 0.565680276
#> [1255] 0.441335285 0.268089259 0.323874444 0.491875280 0.409715069 0.121089069
#> [1261] 0.245798285 0.244245086 0.213179642 0.242113715 0.170215535 0.148002996
#> [1267] 0.212870385 0.127386751 0.208717783 0.258562336 0.156956325 0.303618205
#> [1273] 0.108964191 0.054631101 0.052689230 0.190131803 0.117142083 0.123359521
#> [1279] 0.186485742 0.157120025 0.208994580 0.065854104 0.050634363 0.159028336
#> [1285] 0.012614918 0.086604685 0.025483262 0.031957695 0.081289726 0.015011800
#> [1291] 0.018615196 0.068983059 0.179722797 0.063064638 0.072331974 0.083089627
#> [1297] 0.064228084 0.081713862 0.115182020 0.261043846 0.147236017 0.238885391
#> [1303] 0.215736977 0.225896983 0.251958285 0.267420532 0.204179841 0.291830834
#> [1309] 0.102840130 0.090623848 0.223159729 0.185444822 0.171885802 0.139267989
#> [1315] 0.125794746 0.238305796 0.263807690 0.201119688 0.216645379 0.256772503
#> [1321] 0.184685983 0.071130251 0.178387373 0.273670830 0.261717164 0.249043900
#> [1327] 0.194316451 0.198105873 0.388815692 0.149847271 0.145133072 0.227080077
#> [1333] 0.258739881 0.087022322 0.211926435 0.185357518 0.161262335 0.187941511
#> [1339] 0.054672389 0.101875155 0.024647673 0.219932395 0.038559165 0.108771461
#> [1345] 0.027992317 0.088874713 0.140536383 0.012907212 0.022116042 0.046268637
#> [1351] 0.016174887 0.035595518 0.041925350 0.054071208 0.043079232 0.025733733
#> [1357] 0.017392871 0.027115094 0.029303968 0.031261035 0.007956637 0.136480060
#> [1363] 0.242789366 0.414954371 0.052873160 0.045325584 0.028980049 0.048057755
#> [1369] 0.118832256 0.163097387 0.077536033 0.152236655 0.067968160 0.112090457
#> [1375] 0.163493375 0.100913273 0.128783502 0.287850266 0.208253981 0.207713275
#> [1381] 0.173926091 0.075383982 0.343668479 0.547422638 0.349531425 0.274376502
#> [1387] 0.149643184 0.332707528 0.892202139 0.823656106 0.072109250 0.190267577
#> [1393] 0.145669069 0.119652619 0.395371245 0.246943872 0.095087788 0.110614124
#> [1399] 0.109654762 0.075224941 0.141232353 0.287898242 0.167101368 0.184853092
#> [1405] 0.383320177 0.277890338 0.410912105 0.228158087 0.256690124 0.265656693
#> [1411] 0.126360657 0.076461256 0.282018980 0.233780795 0.197632494 0.318378283
#> [1417] 0.178147622 0.171359740 0.341879497 0.357181015 0.373638173 0.594215156
#> [1423] 0.287058179 0.156960871 0.223331492 0.071210687 0.121579310 0.121068330
#> [1429] 0.152124333 0.138620553 0.045031364 0.020942471 0.033822672 0.095227745
#> [1435] 0.093518866 0.071367703 0.069965646 0.087294566 0.102105429 0.083082530
#> [1441] 0.127355736 0.112794099 0.246861389 0.165227639 0.190465464 0.295434028
#> [1447] 0.265883121 0.257248125 0.243483967 0.305257454 0.273442052 0.166357911
#> [1453] 0.385069639 0.096259693 0.178204164 0.147760400 0.126425020 0.088167149
#> [1459] 0.153056303 0.162068731 0.138043050 0.278474974 0.089179283 0.143387392
#> [1465] 0.185706458 0.125335929 0.248599816 0.149900871 0.104680369 0.053912850
#> [1471] 0.056100421 0.720136374 0.201422263 0.122752180 0.130690401 0.135659113
#> [1477] 0.100492188 0.348714692 0.185338053 0.128645193 0.153276117 0.136519717
#> [1483] 0.298836340 0.265808211 0.170122518 0.182911940 0.200033652 0.544989368
#> [1489] 0.361469335 0.328370408 0.359334932 0.358750789 0.453945932 0.296019192
#> [1495] 0.451297735 0.290383783 0.369532072 0.166921052 0.243504678 0.216859526
#> [1501] 0.189723704 0.185547420 0.211761904 0.134605205 0.267531205 0.259321646
#> [1507] 0.438935476 0.342173005 0.243950344 0.430751326 0.364291997 0.366518688
#> [1513] 0.266978706 0.257947952 0.232204747 0.189400256 0.186241797 0.210785685
#> [1519] 0.228295174 0.251692146 0.268083280 0.187658312 0.185738816 0.221221137
#> [1525] 0.233134473 0.305296757 0.506814044 0.358523203 0.538981382 0.609061379
#> [1531] 0.245494783 0.340617220 0.228672199 0.193371255 0.163818497 0.225341401
#> [1537] 0.221122493 0.356385974 0.481209437 0.348957650 0.233273389 0.266638280
#> [1543] 0.389322976 0.277295255 0.213224576 0.185943389 0.180017591 0.223845871
#> [1549] 0.170178860 0.136392975 0.173062128 0.183985190 0.311226526 0.203861280
#> [1555] 0.183843736 0.177115512 0.448238745 0.338744957 0.454032721 0.186330984
#> [1561] 0.108828881 0.397784490 0.436837726 0.426178155 0.405827844 0.427877258
#> [1567] 0.388243277 0.470890418 0.394095662 0.404788528 0.428427128 0.223625746
#> [1573] 0.237552477 0.235321039 0.325236409 0.299950052 0.234227423 0.269582461
#> [1579] 0.370731135 0.375830643 0.375345997 0.330128315 0.321565231 0.318753939
#> [1585] 0.441373650 0.465292143 0.391821325 0.000000000

# Using dplyr
require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

mutate(bay_race, divergence_score = divergence(white, hispanic,
  asian, black, all_other, population=total_pop))

#> # A tibble: 1,588 x 9
#>    fips     total_pop hispanic white  black  asian all_other county             
#>    <chr>        <dbl>    <dbl> <dbl>  <dbl>  <dbl>     <dbl> <chr>              
#>  1 0600140…      2937   0.0398 0.708 0.0477 0.155     0.0497 Alameda County, Ca…
#>  2 0600140…      1974   0.0765 0.783 0.0157 0.0740    0.0507 Alameda County, Ca…
#>  3 0600140…      4865   0.0820 0.669 0.105  0.0861    0.0573 Alameda County, Ca…
#>  4 0600140…      3703   0.0897 0.655 0.121  0.0729    0.0618 Alameda County, Ca…
#>  5 0600140…      3517   0.0967 0.506 0.265  0.0591    0.0734 Alameda County, Ca…
#>  6 0600140…      1571   0.0802 0.427 0.391  0.0509    0.0503 Alameda County, Ca…
#>  7 0600140…      4206   0.104  0.295 0.492  0.0592    0.0502 Alameda County, Ca…
#>  8 0600140…      3594   0.102  0.316 0.407  0.119     0.0556 Alameda County, Ca…
#>  9 0600140…      2302   0.133  0.310 0.437  0.0556    0.0647 Alameda County, Ca…
#> 10 0600140…      5678   0.142  0.223 0.502  0.0736    0.0594 Alameda County, Ca…
#> # … with 1,578 more rows, and 1 more variable: divergence_score <dbl>

# divergence alsow works with percentages as long as you have
# population totals by observation
bay_race %>%
  mutate_at(vars(hispanic:all_other), list(~(./total_pop))) %>%
  mutate(divergence_score = divergence(white, hispanic, asian,
  black, all_other, population = total_pop))

#> # A tibble: 1,588 x 9
#>    fips    total_pop  hispanic    white    black   asian all_other county       
#>    <chr>       <dbl>     <dbl>    <dbl>    <dbl>   <dbl>     <dbl> <chr>        
#>  1 060014…      2937 0.0000136  2.41e-4  1.62e-5 5.29e-5 0.0000169 Alameda Coun…
#>  2 060014…      1974 0.0000388  3.97e-4  7.96e-6 3.75e-5 0.0000257 Alameda Coun…
#>  3 060014…      4865 0.0000169  1.38e-4  2.16e-5 1.77e-5 0.0000118 Alameda Coun…
#>  4 060014…      3703 0.0000242  1.77e-4  3.27e-5 1.97e-5 0.0000167 Alameda Coun…
#>  5 060014…      3517 0.0000275  1.44e-4  7.54e-5 1.68e-5 0.0000209 Alameda Coun…
#>  6 060014…      1571 0.0000511  2.72e-4  2.49e-4 3.24e-5 0.0000320 Alameda Coun…
#>  7 060014…      4206 0.0000248  7.00e-5  1.17e-4 1.41e-5 0.0000119 Alameda Coun…
#>  8 060014…      3594 0.0000284  8.79e-5  1.13e-4 3.32e-5 0.0000155 Alameda Coun…
#>  9 060014…      2302 0.0000579  1.35e-4  1.90e-4 2.42e-5 0.0000281 Alameda Coun…
#> 10 060014…      5678 0.0000250  3.94e-5  8.83e-5 1.30e-5 0.0000105 Alameda Coun…
#> # … with 1,578 more rows, and 1 more variable: divergence_score <dbl>